### 2021-05-26 #3316 @whedon
[PRE REVIEW]: lattice-symmetries: A package for working with quantum many-body bases
**Submitting author:** @twesterhout (<a href="http://orcid.org/0000-0003-0200-2686">Tom Westerhout</a>)
**Repository:** <a href="https://github.com/twesterhout/lattice-symmetries" target ="_blank">https://github.com/twesterhout/lattice-symmetries</a>
**Version:** v0.8.0
