### 2021-05-27 daaa708
Fix DOIs using betterbib
