+++ paper/paper.bib
@@ -1,114 +1,153 @@
 @article{lauchi2019,
-  title = {$S=\frac{1}{2}$ kagome Heisenberg antiferromagnet revisited},
-  author = {L\"auchli, Andreas M. and Sudan, Julien and Moessner, Roderich},
-  journal = {Phys. Rev. B},
-  volume = {100},
-  issue = {15},
-  pages = {155142},
-  numpages = {7},
-  year = {2019},
-  month = {Oct},
-  publisher = {American Physical Society},
-  doi = {10.1103/PhysRevB.100.155142},
-  url = {https://link.aps.org/doi/10.1103/PhysRevB.100.155142}
+ author = {Läuchli, Andreas M. and Sudan, Julien and Moessner, Roderich},
+ title = {$S=\frac{1}{2}$ kagome {Heisenberg} antiferromagnet revisited},
+ journal = {Phys. Rev. B},
+ volume = {100},
+ issue = {15},
+ pages = {155142},
+ numpages = {7},
+ year = {2019},
+ month = oct,
+ publisher = {American Physical Society (APS)},
+ doi = {10.1103/physrevb.100.155142},
+ url = {https://doi.org/10.1103/physrevb.100.155142},
+ number = {15},
+ source = {Crossref},
+ issn = {2469-9950, 2469-9969},
 }
+
 @article{wietek2018,
-  title = {Sublattice coding algorithm and distributed memory parallelization for large-scale exact diagonalizations of quantum many-body systems},
-  author = {Wietek, Alexander and L\"auchli, Andreas M.},
-  journal = {Phys. Rev. E},
-  volume = {98},
-  issue = {3},
-  pages = {033309},
-  numpages = {10},
-  year = {2018},
-  month = {Sep},
-  publisher = {American Physical Society},
-  doi = {10.1103/PhysRevE.98.033309},
-  url = {https://link.aps.org/doi/10.1103/PhysRevE.98.033309}
+ author = {Wietek, Alexander and Läuchli, Andreas M.},
+ title = {Sublattice coding algorithm and distributed memory parallelization for large-scale exact diagonalizations of quantum many-body systems},
+ journal = {Phys. Rev. E},
+ volume = {98},
+ issue = {3},
+ pages = {033309},
+ numpages = {10},
+ year = {2018},
+ month = sep,
+ publisher = {American Physical Society (APS)},
+ doi = {10.1103/physreve.98.033309},
+ url = {https://doi.org/10.1103/physreve.98.033309},
+ number = {3},
+ source = {Crossref},
+ issn = {2470-0045, 2470-0053},
 }
+
 @article{kawamura2017,
-  title = {Quantum lattice model solver $\text{H}\Phi$},
-  journal = {Computer Physics Communications},
-  volume = {217},
-  pages = {180-192},
-  year = {2017},
-  issn = {0010-4655},
-  doi = {https://doi.org/10.1016/j.cpc.2017.04.006},
-  url = {https://www.sciencedirect.com/science/article/pii/S0010465517301200},
-  author = {Mitsuaki Kawamura and Kazuyoshi Yoshimi and Takahiro Misawa and Youhei Yamaji and Synge Todo and Naoki Kawashima},
+ author = {Kawamura, Mitsuaki and Yoshimi, Kazuyoshi and Misawa, Takahiro and Yamaji, Youhei and Todo, Synge and Kawashima, Naoki},
+ title = {Quantum lattice model solver {H$\Phi$}},
+ journal = {Comput. Phys. Commun.},
+ volume = {217},
+ pages = {180-192},
+ year = {2017},
+ issn = {0010-4655},
+ doi = {10.1016/j.cpc.2017.04.006},
+ url = {https://doi.org/10.1016/j.cpc.2017.04.006},
[…]
