### 2021-05-27 #3316 @twesterhout
I've cleaned up the `paper.bib` file using [betterbib](https://github.com/nschloe/betterbib) (see [this commit](https://github.com/twesterhout/lattice-symmetries/commit/daaa7084aaf59401a7c566947e494de7b4cde461)). DOIs should be okay now.

Regarding potential reviewers, I would suggest nmoran and mdavezac.
