### 2021-06-28 #3316 @jedbrown
Not a big deal, but you've got some warnings.

gcc-11:

```
/home/jed/joss/lattice-symmetries/src/operator.cpp: In instantiation of ‘lattice_symmetries::block_acc_t<T>::block_acc_t(uint64_t) [with T = std::complex<float>; uint64_t = long unsigned int]’:
/home/jed/joss/lattice-symmetries/src/operator.cpp:517:33:   required from ‘outcome_v2_1e90094c::result<void> lattice_symmetries::matmat_helper(const ls_operator&, uint64_t, uint64_t, const T*, uint64_t, T*, uint64_t) [with T = std::complex<float>; outcome_v2_1e90094c::result<void> = outcome_v2_1e90094c::basic_result<void, std::error_code, outcome_v2_1e90094c::policy::error_code_throw_as_system_error<void, std::error_code, void> >; ls_operator = ls_operator; uint64_t = long unsigned int]’
/home/jed/joss/lattice-symmetries/src/operator.cpp:690:35:   required from here
/home/jed/joss/lattice-symmetries/src/operator.cpp:457:20: warning: ‘void* memset(void*, int, size_t)’ clearing an object of non-trivial type ‘class std::complex<double>’; use assignment or value-initialization instead [-Wclass-memaccess]
  457 |         std::memset(data.get(), 0, sizeof(acc_t) * num_threads * stride);
      |         ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /home/jed/joss/lattice-symmetries/include/lattice_symmetries/lattice_symmetries.h:93,
                 from /home/jed/joss/lattice-symmetries/src/symmetry.hpp:31,
                 from /home/jed/joss/lattice-symmetries/src/basis.hpp:32,
                 from /home/jed/joss/lattice-symmetries/src/operator.cpp:2:
/usr/include/c++/11.1.0/complex:1229:11: note: ‘class std::complex<double>’ declared here
 1229 |     class complex<double>
      |           ^~~~~~~~~~~~~~~
/home/jed/joss/lattice-symmetries/src/operator.cpp: In instantiation of ‘void lattice_symmetries::block_acc_t<T>::set_zero(unsigned int) const [with T = std::complex<float>]’:
/home/jed/joss/lattice-symmetries/src/operator.cpp:536:27:   required from ‘outcome_v2_1e90094c::result<void> lattice_symmetries::matmat_helper(const ls_operator&, uint64_t, uint64_t, const T*, uint64_t, T*, uint64_t) [with T = std::complex<float>; outcome_v2_1e90094c::result<void> = outcome_v2_1e90094c::basic_result<void, std::error_code, outcome_v2_1e90094c::policy::error_code_throw_as_system_error<void, std::error_code, void> >; ls_operator = ls_operator; uint64_t = long unsigned int]’
/home/jed/joss/lattice-symmetries/src/operator.cpp:690:35:   required from here
/home/jed/joss/lattice-symmetries/src/operator.cpp:468:20: warning: ‘void* memset(void*, int, size_t)’ clearing an object of non-trivial type ‘class std::complex<double>’; use assignment or value-initialization instead [-Wclass-memaccess]
  468 |         std::memset(data.get() + stride * thread_num, 0, sizeof(acc_t) * block_size);
      |         ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /home/jed/joss/lattice-symmetries/include/lattice_symmetries/lattice_symmetries.h:93,
                 from /home/jed/joss/lattice-symmetries/src/symmetry.hpp:31,
                 from /home/jed/joss/lattice-symmetries/src/basis.hpp:32,
                 from /home/jed/joss/lattice-symmetries/src/operator.cpp:2:
/usr/include/c++/11.1.0/complex:1229:11: note: ‘class std::complex<double>’ declared here
 1229 |     class complex<double>
      |           ^~~~~~~~~~~~~~~
/home/jed/joss/lattice-symmetries/src/operator.cpp: In instantiation of ‘lattice_symmetries::block_acc_t<T>::block_acc_t(uint64_t) [with T = std::complex<double>; uint64_t = long unsigned int]’:
/home/jed/joss/lattice-symmetries/src/operator.cpp:517:33:   required from ‘outcome_v2_1e90094c::result<void> lattice_symmetries::matmat_helper(const ls_operator&, uint64_t, uint64_t, const T*, uint64_t, T*, uint64_t) [with T = std::complex<double>; outcome_v2_1e90094c::result<void> = outcome_v2_1e90094c::basic_result<void, std::error_code, outcome_v2_1e90094c::policy::error_code_throw_as_system_error<void, std::error_code, void> >; ls_operator = ls_operator; uint64_t = long unsigned int]’
/home/jed/joss/lattice-symmetries/src/operator.cpp:691:36:   required from here
/home/jed/joss/lattice-symmetries/src/operator.cpp:457:20: warning: ‘void* memset(void*, int, size_t)’ clearing an object of non-trivial type ‘class std::complex<double>’; use assignment or value-initialization instead [-Wclass-memaccess]
  457 |         std::memset(data.get(), 0, sizeof(acc_t) * num_threads * stride);
      |         ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /home/jed/joss/lattice-symmetries/include/lattice_symmetries/lattice_symmetries.h:93,
                 from /home/jed/joss/lattice-symmetries/src/symmetry.hpp:31,
                 from /home/jed/joss/lattice-symmetries/src/basis.hpp:32,
                 from /home/jed/joss/lattice-symmetries/src/operator.cpp:2:
/usr/include/c++/11.1.0/complex:1229:11: note: ‘class std::complex<double>’ declared here
 1229 |     class complex<double>
      |           ^~~~~~~~~~~~~~~
/home/jed/joss/lattice-symmetries/src/operator.cpp: In instantiation of ‘void lattice_symmetries::block_acc_t<T>::set_zero(unsigned int) const [with T = std::complex<double>]’:
/home/jed/joss/lattice-symmetries/src/operator.cpp:536:27:   required from ‘outcome_v2_1e90094c::result<void> lattice_symmetries::matmat_helper(const ls_operator&, uint64_t, uint64_t, const T*, uint64_t, T*, uint64_t) [with T = std::complex<double>; outcome_v2_1e90094c::result<void> = outcome_v2_1e90094c::basic_result<void, std::error_code, outcome_v2_1e90094c::policy::error_code_throw_as_system_error<void, std::error_code, void> >; ls_operator = ls_operator; uint64_t = long unsigned int]’
/home/jed/joss/lattice-symmetries/src/operator.cpp:691:36:   required from here
/home/jed/joss/lattice-symmetries/src/operator.cpp:468:20: warning: ‘void* memset(void*, int, size_t)’ clearing an object of non-trivial type ‘class std::complex<double>’; use assignment or value-initialization instead [-Wclass-memaccess]
  468 |         std::memset(data.get() + stride * thread_num, 0, sizeof(acc_t) * block_size);
      |         ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /home/jed/joss/lattice-symmetries/include/lattice_symmetries/lattice_symmetries.h:93,
                 from /home/jed/joss/lattice-symmetries/src/symmetry.hpp:31,
                 from /home/jed/joss/lattice-symmetries/src/basis.hpp:32,
                 from /home/jed/joss/lattice-symmetries/src/operator.cpp:2:
/usr/include/c++/11.1.0/complex:1229:11: note: ‘class std::complex<double>’ declared here
 1229 |     class complex<double>
      |           ^~~~~~~~~~~~~~~
```

clang-12:

```
[ 22%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx2.dir/src/cpu/benes_forward_64.cpp.o
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:104:20: warning: unused variable 'cache_line_mask' [-Wunused-variable]
    constexpr auto cache_line_mask         = (~uint64_t{0}) << 6U;
                   ^
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:138:78: warning: implicit conversion loses integer precision: 'uint64_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
    return static_cast<uint64_t>(data - original_data) + linear_search(data, size, key_v);
                                                         ~~~~~~~~~~~~~       ^~~~
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:104:20: warning: unused variable 'cache_line_mask' [-Wunused-variable]
    constexpr auto cache_line_mask         = (~uint64_t{0}) << 6U;
                   ^
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:138:78: warning: implicit conversion loses integer precision: 'uint64_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
    return static_cast<uint64_t>(data - original_data) + linear_search(data, size, key_v);
                                                         ~~~~~~~~~~~~~       ^~~~
2 warnings generated.
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:104:20: warning: unused variable 'cache_line_mask' [-Wunused-variable]
    constexpr auto cache_line_mask         = (~uint64_t{0}) << 6U;
                   ^
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:138:78: warning: implicit conversion loses integer precision: 'uint64_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
    return static_cast<uint64_t>(data - original_data) + linear_search(data, size, key_v);
                                                         ~~~~~~~~~~~~~       ^~~~
2 warnings generated.
[ 29%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse2.dir/src/cpu/benes_forward_64.cpp.o
[ 29%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse4.dir/src/cpu/benes_forward_64.cpp.o
2 warnings generated.
[ 32%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse2.dir/src/cpu/benes_forward_512.cpp.o
[ 35%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx2.dir/src/cpu/benes_forward_512.cpp.o
[ 38%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx2.dir/src/cpu/state_info.cpp.o
[ 41%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse2.dir/src/cpu/state_info.cpp.o
[ 45%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse4.dir/src/cpu/benes_forward_512.cpp.o
[ 48%] Building CXX object CMakeFiles/lattice_symmetries_kernels_sse4.dir/src/cpu/state_info.cpp.o
[ 51%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx.dir/src/cpu/search_sorted.cpp.o
[ 51%] Built target lattice_symmetries_kernels_avx2
[ 54%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx.dir/src/cpu/benes_forward_64.cpp.o
[ 54%] Built target lattice_symmetries_kernels_sse4
[ 58%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx.dir/src/cpu/benes_forward_512.cpp.o
[ 58%] Built target lattice_symmetries_kernels_sse2
[ 61%] Building CXX object CMakeFiles/lattice_symmetries_kernels_avx.dir/src/cpu/state_info.cpp.o
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:104:20: warning: unused variable 'cache_line_mask' [-Wunused-variable]
    constexpr auto cache_line_mask         = (~uint64_t{0}) << 6U;
                   ^
/home/jed/joss/lattice-symmetries/src/cpu/search_sorted.cpp:138:78: warning: implicit conversion loses integer precision: 'uint64_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
    return static_cast<uint64_t>(data - original_data) + linear_search(data, size, key_v);
                                                         ~~~~~~~~~~~~~       ^~~~
2 warnings generated.
[ 61%] Built target lattice_symmetries_kernels_avx
[ 74%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/basis.cpp.o
[ 74%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/group.cpp.o
[ 74%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/error_handling.cpp.o
[ 74%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/cache.cpp.o
[ 77%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/network.cpp.o
/home/jed/joss/lattice-symmetries/src/cache.cpp:96:27: warning: unused variable 'mask' [-Wunused-variable]
        auto const        mask             = size - 1;
                          ^
/home/jed/joss/lattice-symmetries/src/cache.cpp:368:22: warning: implicit conversion changes signedness: 'long' to 'uint64_t' (aka 'unsigned long') [-Wsign-conversion]
        *out = index - _states.data();
             ~ ~~~~~~^~~~~~~~~~~~~~~~
/home/jed/joss/lattice-symmetries/src/group.cpp:175:35: warning: implicit conversion loses integer precision: 'unsigned int' to 'std::vector<unsigned short>::value_type' (aka 'unsigned short') [-Wimplicit-int-conversion]
            permutation.push_back(i);
                        ~~~~~~~~~ ^
2 warnings generated.
[ 80%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/operator.cpp.o
1 warning generated.
[ 83%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/permutation.cpp.o
[ 87%] Building CXX object CMakeFiles/lattice_symmetries.dir/src/symmetry.cpp.o
/home/jed/joss/lattice-symmetries/src/network.cpp:34:21: warning: implicit conversion loses integer precision: 'unsigned int' to 'uint16_t' (aka 'unsigned short') [-Wimplicit-int-conversion]
        deltas[i] = d;
                  ~ ^
1 warning generated.
[ 90%] Linking CXX static library liblattice_symmetries.a
[ 90%] Built target lattice_symmetries
[ 96%] Building CXX object test/CMakeFiles/LatticeSymmetriesTests.dir/src/test_api.cpp.o
[ 96%] Building CXX object test/CMakeFiles/LatticeSymmetriesTests.dir/src/test_main.cpp.o
/home/jed/joss/lattice-symmetries/test/src/test_api.cpp:35:1: warning: non-void function does not return a value [-Wreturn-type]
}
^
1 warning generated.
```
