### 2021-06-28 750c57a
Fix warnings for gcc-10
