### 2021-06-28 c0205bd
Fix warnings in clang-11
