### 2021-07-05 b388808
Other may be reallocated and then s becomes invalid
