### 2021-07-06 eaf578f
Fix a bug when spin_inversion==-1 is not taken into account for the zeroth group element
