### 2021-07-24 #3316 @jedbrown
@joselado @bastikr :wave: We're searching for an additional reviewer for this submission to JOSS, the [Journal of Open Source Software](https://joss.theoj.org/about), and it looks like you've done relevant work in this area. Our [review process]((https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)) is unique and collaborative, conducted within GitHub issues. Would you be interested and available to review this submission?
