### 2021-07-26 twesterhout/lattice-symmetries#1 @jeffhammond
generic/portable support for non-x86 systems
I am reviewing https://github.com/openjournals/joss-reviews/issues/3537 and was unable to test on my Apple M1 (my machine is not setup to compiler for x86 even if Rosetta2 supports executing such binaries).

I see that the code has support for all the x86 SIMD instruction sets via Agner Fog's library, but there does not appear to be a generic scalar implementation that would allow the code to be used on any platform with a ISO C++ compiler.

If it was completely obvious to me, I would contribute this feature but I do not think I am qualified to unwind this dependency and implement generic platform support.
