### 2021-07-26 #3537 @jeffhammond
# Summary

This paper describes a software package for exact diagonalization of lattice Hamiltonians. This is an important domain of computational science.  The software for solving problems in this space is challenging to write and use.  The author's contribution here is significant, and I have personal experience with how important it is to have efficient implementations of this capability.

This submission seems like an appropriate and useful contribution to JOSS and I recommend that it be accepted.

# Shortcomings of this review

I did not attempt to verify the performance claims in the paper.  I did not download or install QuSpin or SPINPACK.  On the other hand, I downloaded and verified SpinED, in addition to lattice-symmetries, and was satisfied with their compliance with the documentation and reported behavior.

# Potential improvements

- "It allows one to treat systems of up to O(42) sites on a single node" is a common but incorrect use of big-O notation.  O(42) is O(1).  It seems that the number 42 is sufficient here, or if there is a possibility that 44 or 45 might also be possible, then using something like "at least 42" might be appropriate.
- The requirement of an x86 process is documented in the repo but not the JOSS manuscript.  I would like the latter to be more explicit about this.  It would be nice - but is not necessary - to note that the project makes use of Agner Fog's library for SIMD support (https://github.com/vectorclass/version2).
- There is no explicit contributor guide, although the necessary information is obvious enough on the GitHub page.
