### 2021-07-27 twesterhout/lattice-symmetries#1 @twesterhout
@jeffhammond thank you for spending time to review this package!

Since most compute clusters use CPUs which are based on x86_64, I never even considered supporting other architectures. Writing a high-performance arm implementation would be too much work for too little benefit (since publication-quality results will be produced on clusters anyway), but I like the idea of a scalar fallback implementation for easy testing.

I'm happy to give scalar implementation a go, but I have no arm/aarch64 machine to test it on. Would you be willing to work together on this and use your Apple M1 as a guinea pig for compiling the code?
