### 2021-07-27 twesterhout/lattice-symmetries#1 @jeffhammond
Yes, I would be happy to test a generic port of this code on all of my ARM systems.  I have lots of them 😎

Below is more info about ARM adoption than you ever wanted to know...

You can buy a 64-bit [Raspberry Pi 4](https://www.raspberrypi.org/products/raspberry-pi-4-model-b/) system for less than €50.  They are not the most powerful systems but they were just fine for porting many codes.  I have developed LLVM and NWChem on Raspberry Pi or similar systems, both of which are millions of lines of code.

AWS offers 64-core [ARM servers](https://aws.amazon.com/ec2/graviton/) that are competitive with the best x86 CPUs for both compute- and memory-bound workloads, and are, at the very least, useful for doing ARM porting if you don't want to own hardware locally.

Ampere is building ARM servers for data centers, which are more powerful than many x86 servers ([details](https://www.anandtech.com/show/16315/the-ampere-altra-review)).  Oracle offers Ampere servers in their [cloud](https://www.oracle.com/cloud/compute/arm/) right now.  I recently [presented](https://youtu.be/xq_sj4nAk3k) on the performance of these CPUs versus Intel Xeon processors for the NWChem quantum chemistry workload.

The current fastest supercomputer in the world, Fugaku, is an ARM system. The next two after that are Power9+NVIDIA.  Number 4 is a custom non-x86 chip.  At #5, we see the first x86 supercomputer, which gets a large portion of its performance from NVIDIA GPUs.  [Top500](https://top500.org/lists/top500/2021/06/) has details.

I can help you get access to the same CPUs as are found in Fugaku at [Bristol](https://gw4.ac.uk/isambard/) or [Georgia Tech](http://crnch.gatech.edu/rg-arm-hpc) if you like.  I guess you in Europe so the ping times to Bristol should be better.

In time, you will see an increasing number of ARM-based HPC systems in the world, such as those based on the [NVIDIA Grace](https://www.nvidia.com/en-us/data-center/grace-cpu/) processor, which are already planned at [Los Alamos](https://www.lanl.gov/discover/news-release-archive/2021/April/0412-nvidia.php) and [CSCS](https://nvidianews.nvidia.com/news/swiss-national-supercomputing-centre-hewlett-packard-enterprise-and-nvidia-announce-worlds-most-powerful-ai-capable-supercomputer).

I didn't even mention what Apple, Qualcomm and SiPearl are doing with ARM right now, but the internet has details.

Anyways, I work for NVIDIA and building ARM HPC software is a big part of my day job, so I'm biased, but even when I worked for Intel (2014-2021), I did not encourage people to write x86-specific code.
