### 2021-07-28 twesterhout/lattice-symmetries#1 @twesterhout
Wow! This is a lot of very good info. I wasn't aware of these developments... Thank you so much!

This argument is so convincing that now I want a fast ARM implementation :D And then compare the performance of Ampere to AMD EPYC (which I've used for benchmarking).

The easiest is to swap [vectorclass/version2](https://github.com/vectorclass/version2) for e.g. [simd-everywhere/simde](https://github.com/simd-everywhere/simde) and play around with CMake to make it compile on ARM as well.

A completely different solution is to use [Halide](https://halide-lang.org/) to generate kernels for various platforms and then do run-time dispatch similarly to how it's done now. The upside is that performance tuning for a few specific architectures becomes way easier, and it would be trivial to get a proof-of-concept GPU version running. The downside is that it's much more work, and I'm not entirely sure that the performance will be on par with the current version.

So I'm still thinking about the best way to approach this issue...
