### 2021-07-29 twesterhout/lattice-symmetries#1 @jeffhammond
Yeah, SIMDe is one of the recommended solutions here.  An ARM friend mentioned he ported most of vectorclass to ARM with it, but there were a few small gaps (that might not affect you).

Halide and other DSLs are very interesting.  I looked at Halide a bit in the past but the learning curve was steep.  I will have to look at your code more to see what other options exist.  If there's a massive amount of task or data parallelism, there are many options.  If one has to manage SIMD/SIMT and L1/SLMs explicitly, it's harder.
