### 2021-07-29 twesterhout/lattice-symmetries#1 @jlinford
Hi!  Porting vectorclass to Arm (Gravition 2 actually) wasn't too hard.   SIMDe's not a drop-in replacement (see below) but it gets you nearly there.  Most of the code changes were due to differences in the compilers between systems; you'd see the same sort of issues moving from x86 to POWER.  Someone comfortable writing AVX512 intrinsics is also comfortable leveraging Intel compiler intrinsics and inline x86 assembly, and SIMDe can't help with that.

I would *love* to send you a pull request, but Arm's lawyers would get upset since I don't yet have the proper paperwork in place to contribute to this open source project.  While I'm getting that sorted, I can explain the changes and let you take it from there, if you like.

### To get started:
In version2/instrset.h:
```
//#include <x86intrin.h>                 // Gcc or Clang compiler
#include <simde/x86/avx512.h>
```

### Problems and solutions:
* vectorclass assumed the signedness of char, e.g. `../version2/vectori128.h:1037:45: error: narrowing conversion of ‘-1’ from ‘int’ to ‘char’ inside { } [-Wnarrowing]`.  Either specify `signed char` in the appropriate place, or compile with `-fsigned-char`.

* vectorclass macros don't always paren-protect their arguments, e.g. `../version2/vectorf128.h:2126:82: error: macro "_mm_castsi128_pd" passed 4 arguments, but takes just 1`.  Fixed by use double parens, e.g. `_mm_castsi128_pd((args))`

* vectorclass assumes your compiler supports Intel compiler intrinsics, e.g. `../version2/instrset.h:289:22: error: ‘_mm_popcnt_u32’ was not declared in this scope`.  Fixed by rewriting the code to not use Intel intrinsics.

* vectorclass uses inline x86 assembly, e.g. `__asm("bsrl %1, %0" : "=r"(r) : "r"(a) : );`.  Fixed by rewriting the code to use GCC builtin functions.

* vectorclass doesn't know `__aarch64__` is a 64-bit platform.  Fixed by defining `__x86_64__` when `__aarch64__` is defined.  Equating `__x86_64__` with `__aarch64__` is a dirty hack and shouldn't be done, but it's a first step.

All this took less than an afternoon.

Some tests fail due to caveats.  See https://github.com/simd-everywhere/simde#caveats.  I've attached the output from `./runtest.sh test1.lst` so you can see which tests are failing.

[test1.out.gz](https://github.com/twesterhout/lattice-symmetries/files/6901059/test1.out.gz)

Stats:

Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz:
5490/5490 tests executed successfully in 62 minutes 5 seconds

Graviton 2:
5297/5490 tests executed successfully in 135 minutes 0 seconds
