### 2021-08-09 #3537 @jeffhammond
My verdict in https://github.com/openjournals/joss-reviews/issues/3537#issuecomment-886594757 still stands ("I recommend that it be accepted").
