### 2021-08-09 #3537 @joselado
# Summary #
This software presents a high-performance implementation of spin operators in a quantum many-body context. High-performance solvers for quantum many-body Hamiltonians are a tool of critical importance in quantum physics and condensed matter physics, and as a result, the current implementation is of very high value for the computational quantum community. The code is well documented and easy to use, and the manuscript presents a comprehensive summary of the computational package. Given all the points above, I recommend the manuscript to be accepted in JOSS.

# Suggestions #
I would suggest that it can be interesting to write in the manuscript an example formula of the Hamiltonians that the code can solve. This is, of course, not necessary for experts, yet it can be informative to potential readers and users not familiar with the topic.

# Recommendation #
As noted above, I believe that this software is a highly valuable contribution to the community and fulfills all the acceptance criteria of JOSS. Therefore, I strongly recommend its publication in JOSS.
