### 2021-08-09 #3537 @jedbrown
Thanks for your reviews, @jeffhammond and @joselado.

@jeffhammond Could you please tick the "performance" box in your review checklist if you feel it has reasonably been addressed. (I see your review adds some nuance that you didn't compare with every alternative, and that's entirely fine.)

@twesterhout Could you please revise based on reviewer suggestions and let me know when ready?
