### 2021-08-09 4223169
Add functions to dump Benes network masks (for external use)
