+++ python/test/test_basis.py
@@ -65,6 +65,89 @@ def test_index():
     # evals, evecs = ls.diagonalize(hamiltonian)
     # print(evals)
 
+def test_construction():
+    symmetries = [
+        # ls.Symmetry([18, 19,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14, 15, 16, 17], sector=5),
+        # ls.Symmetry([19, 18, 17, 16, 15, 14, 13, 12, 11, 10,  9,  8,  7,  6,  5,  4,  3,  2,  1, 0], sector=0)
+    ]
+    basis = ls.SpinBasis(ls.Group(symmetries), number_spins=20, hamming_weight=10, spin_inversion=None)
+    basis.build()
 
-test_index()
+    interactions = [
+        ls.Interaction([[0.25, 0, 0, 0], [0, -0.25, 0.5, 0], [0,  0.5 , -0.25,  0.  ], [ 0.  ,  0.
+            ,  0.  ,  0.25]], [[0, 1], [2, 3], [4, 5], [6, 7], [8, 9], [10, 11], [12, 13], [14, 15],
+                [16, 17], [18, 19]]),
+
+        ls.Interaction([[ 0.25,  0.  ,  0.  ,  0.  ],
+               [ 0.  , -0.25,  0.5 ,  0.  ],
+               [ 0.  ,  0.5 , -0.25,  0.  ],
+               [ 0.  ,  0.  ,  0.  ,  0.25]],
+            [[0, 2], [1, 3], [2, 4], [3, 5], [4, 6], [5, 7], [6, 8], [7, 9], [8, 10], [9, 11], [10, 12], [11, 13], [12, 14], [13, 15], [14, 16], [15, 17], [16, 18], [17, 19], [18, 0], [19, 1]]
+        ),
+        ls.Interaction([[-0.0625, -0.    , -0.    , -0.    ],
+               [-0.    ,  0.0625, -0.125 , -0.    ],
+               [-0.    , -0.125 ,  0.0625, -0.    ],
+               [-0.    , -0.    , -0.    , -0.0625]],
+            [[0, 4], [1, 5], [2, 6], [3, 7], [4, 8], [5, 9], [6, 10], [7, 11], [8, 12], [9, 13], [10, 14], [11, 15], [12, 16], [13, 17], [14, 18], [15, 19], [16, 0], [17, 1], [18, 2], [19, 3]]
+        ),
+        ls.Interaction([[ 0.02777778,  0.        ,  0.        ,  0.        ],
+               [ 0.        , -0.02777778,  0.05555556,  0.        ],
+               [ 0.        ,  0.05555556, -0.02777778,  0.        ],
+               [ 0.        ,  0.        ,  0.        ,  0.02777778]],
+            [[0, 6], [1, 7], [2, 8], [3, 9], [4, 10], [5, 11], [6, 12], [7, 13], [8, 14], [9, 15],
+                [10, 16], [11, 17], [12, 18], [13, 19], [14, 0], [15, 1], [16, 2], [17, 3], [18, 4],
+                [19, 5]]),
+        ls.Interaction([[-0.015625, -0.      , -0.      , -0.      ],
+               [-0.      ,  0.015625, -0.03125 , -0.      ],
+               [-0.      , -0.03125 ,  0.015625, -0.      ],
+               [-0.      , -0.      , -0.      , -0.015625]],
+            [[0, 8], [1, 9], [2, 10], [3, 11], [4, 12], [5, 13], [6, 14], [7, 15], [8, 16], [9, 17],
[…]
