### 2021-08-10 twesterhout/lattice-symmetries#1 @twesterhout
Sorry for the delay, I was moving and had to take a few days off.

@jlinford thank you for your insights, they're very helpful!

I still haven't decided whether Halide or SIMDe is the right way to go. In lattice-symmetries, SIMD is currently used for two things:

1. Computing representatives of basis configurations. This amounts to a bunch of bit manipulation intrinsics and multiple reductions.
2. Searching for representatives, i.e. we have a vectorized binary search.

I gave 1. a try in Halide (you can find some work-in-progress code here https://github.com/twesterhout/lattice-symmetries-halide). It seems to produce reasonable C code, but I haven't analyzed assembly yet. The reason why I'd like to use Halide here is that it's trivial to change how loops are unrolled and vectorized, and I'm afraid that simply processing 512 bits at a time (what the current implementation does) is sub-optimal when natural vector size is 128 bits (i.e. I see that gcc dumps stuff to the stack rather than doing everything in registers).
My plan now is to benchmark and optimize this Halide implementation and see how fast I can get it to work.

2. definitely needs to be implemented in SIMDe, but I think it's not too difficult.

Everything else is pure C & C++ code and it should just compile for other architectures. If I manage to get Halide implementation of 1. to perform reasonably well, I'll generate ARM kernels and write some wrapper code and ping you @jeffhammond to try compiling it on one of your machines.
