### 2021-08-12 e60a315
Expose shifts as well (we need it for Halide)
