### 2021-08-12 06838f6
Incorporate Jeff Hammond's feedback. Remove O(42) in favor of 'at least 42'. Add a reference to VCL. Add a note about x86-64 compatibility.
