+++ paper/paper.bib
@@ -151,3 +151,23 @@
  year = {2017},
  url = {\url{https://www-e.ovgu.de/jschulen/spin/}},
 }
+
+@misc{vectorclass,
+  author = {Fog, Agner},
+  title = {{VCL:} {C++} vector class library},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/vectorclass/version2}},
+  commit = {46a58414bc6128b61d77186656860ca7b0b961d8},
+}
+
+@misc{amd1999,
+  author = {{AMD Corporation}},
+  title = {{AMD} Discloses New Technologies at {Microprocessor Forum} --- Including x86 64-bit Architecture and {Lightning Data Transport}$^{{\sc TM}}$},
+  howpublished = {Press release},
+  day = {5},
+  month = oct,
+  year = {1999},
+  url = {\url{http://www.amd.com/news/prodpr/99105.html}},
+}
