### 2021-08-12 ff35f4f
Incorporate Jose Lado's feedback. Add a sentence about supported operators and provide a few example formulas of well-known models
