### 2021-08-12 #3537 @twesterhout
Thank you @jeffhammond and @joselado for your reviews! I have modified the manuscript according to your suggestions. Please, see twesterhout/lattice-symmetries@06838f62b2b9ece183dadb8dc007f6b71b903d0c and twesterhout/lattice-symmetries@ff35f4f8c54917678e536ec169522ec797cf426a commits.

Please, let me know if anything else needs to be corrected/improved.
