### 2021-08-12 #3537 @joselado
Thank you @twesterhout for the updated manuscript. The current version looks perfect, I recommend publishing the work as it is in JOSS.
