+++ paper/paper.bib
@@ -162,6 +162,15 @@
   commit = {46a58414bc6128b61d77186656860ca7b0b961d8},
 }
 
+@misc{qsl2021,
+  author = {Astrakhantsev, Nikita},
+  title = {QSL\_at\_QC},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/nikita-astronaut/QSL\_at\_QC}},
+}
+
 @misc{amd1999,
   author = {{AMD Corporation}},
   title = {{AMD} Discloses New Technologies at {Microprocessor Forum} --- Including x86 64-bit Architecture and {Lightning Data Transport}$^{{\sc TM}}$},
