### 2021-08-16 #3537 @twesterhout
One final tiny correction: I've added a link to Github repo of the project which uses lattice-symmetries to simulate quantum circuits (twesterhout/lattice-symmetries@99254416b9edb2bbf973cac98fcc9d1063e87faf). I haven't done it earlier because the code only just became public.

@jedbrown are there any further steps I need to take?
