### 2021-08-17 0fd27f8
Use long journal names (generated with betterbib)
