+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{lauchi2019,
  author = {Läuchli, Andreas M. and Sudan, Julien and Moessner, Roderich},
  title = {$S=\frac{1}{2}$ kagome {Heisenberg} antiferromagnet revisited},
- journal = {Phys. Rev. B},
+ journal = {Physical Review B},
  volume = {100},
  issue = {15},
  pages = {155142},
@@ -19,7 +19,7 @@
 @article{wietek2018,
  author = {Wietek, Alexander and Läuchli, Andreas M.},
  title = {Sublattice coding algorithm and distributed memory parallelization for large-scale exact diagonalizations of quantum many-body systems},
- journal = {Phys. Rev. E},
+ journal = {Physical Review E},
  volume = {98},
  issue = {3},
  pages = {033309},
@@ -37,7 +37,7 @@
 @article{kawamura2017,
  author = {Kawamura, Mitsuaki and Yoshimi, Kazuyoshi and Misawa, Takahiro and Yamaji, Youhei and Todo, Synge and Kawashima, Naoki},
  title = {Quantum lattice model solver {H$\Phi$}},
- journal = {Comput. Phys. Commun.},
+ journal = {Computer Physics Communications},
  volume = {217},
  pages = {180-192},
  year = {2017},
@@ -51,8 +51,8 @@
 
 @article{weinberg2017,
  author = {Weinberg, Phillip and Bukov, Marin},
- title = {{QuSpin:} {A} {Python} package for dynamics and exact diagonalisation of quantum many body systems part {I}: {Spin} chains},
- journal = {SciPost Phys.},
+ title = {{QuSpin:} {A} {Python} package for dynamics and exact diagonalisation of quantum many body systems part I: {Spin} chains},
+ journal = {SciPost Physics},
  volume = {2},
  issue = {1},
  pages = {003},
@@ -101,8 +101,8 @@
  author = {Astrakhantsev, Nikita and Westerhout, Tom and Tiwari, Apoorv and Choo, Kenny and Chen, Ao and Fischer, Mark H and Carleo, Giuseppe and Neupert, Titus},
  title = {Broken-Symmetry Ground States of the {Heisenberg} model on the Pyrochlore Lattice},
  journal = {arXiv Condensed Matter},
- archivePrefix={arXiv},
- eprint={2101.08787},
+ archiveprefix = {arXiv},
+ eprint = {2101.08787},
  year = {2021},
 }
 
@@ -110,15 +110,15 @@
  author = {Bagrov, Andrey A and Iliasov, Askar A and Westerhout, Tom},
  title = {Kinetic samplers for neural quantum states},
  journal = {arXiv Condensed Matter},
- archivePrefix={arXiv},
- eprint={2011.02986},
+ archiveprefix = {arXiv},
+ eprint = {2011.02986},
  year = {2020},
 }
 
 @article{westerhout2020,
  author = {Westerhout, Tom and Astrakhantsev, Nikita and Tikhonov, Konstantin S. and Katsnelson, Mikhail I. and Bagrov, Andrey A.},
  title = {Generalization properties of neural network approximations to frustrated magnet ground states},
- journal = {Nat Commun},
+ journal = {Nature Communications},
  volume = {11},
  number = {1},
  pages = {1--8},
@@ -133,8 +133,8 @@
 
 @article{stathopoulos2010,
  author = {Stathopoulos, Andreas and McCombs, James R.},
- title = {Primme},
- journal = {ACM Trans. Math. Softw.},
+ title = {{PRIMME}},
+ journal = {ACM Transactions on Mathematical Software},
  volume = {37},
  number = {2},
[…]
