+++ paper/paper.bib
@@ -51,7 +51,7 @@
 
 @article{weinberg2017,
  author = {Weinberg, Phillip and Bukov, Marin},
- title = {{QuSpin:} {A} {Python} package for dynamics and exact diagonalisation of quantum many body systems part I: {Spin} chains},
+ title = {{QuSpin:} {A} {Python} package for dynamics and exact diagonalisation of quantum many body systems part {I}: {Spin} chains},
  journal = {SciPost Physics},
  volume = {2},
  issue = {1},
@@ -79,7 +79,7 @@
  year = {2020},
  publisher = {GitHub},
  journal = {GitHub repository},
- howpublished = {\url {https://github.com/twesterhout/spin-ed}},
+ howpublished = {\url{https://github.com/twesterhout/spin-ed}},
 }
 
 @inproceedings{sandvik2010,
@@ -133,7 +133,7 @@
 
 @article{stathopoulos2010,
  author = {Stathopoulos, Andreas and McCombs, James R.},
- title = {{PRIMME}},
+ title = {{PRIMME}: preconditioned iterative multimethod eigensolver --- methods and software description},
  journal = {ACM Transactions on Mathematical Software},
  volume = {37},
  number = {2},
@@ -143,7 +143,6 @@
  doi = {10.1145/1731022.1731031},
  source = {Crossref},
  url = {https://doi.org/10.1145/1731022.1731031},
- subtitle = {preconditioned iterative multimethod eigensolver---methods and software description},
  issn = {0098-3500, 1557-7295},
  month = apr,
 }
@@ -162,7 +161,7 @@
  year = {2020},
  publisher = {GitHub},
  journal = {GitHub repository},
- howpublished = {\url {https://github.com/vectorclass/version2}},
+ howpublished = {\url{https://github.com/vectorclass/version2}},
  commit = {46a58414bc6128b61d77186656860ca7b0b961d8},
 }
 
@@ -172,15 +171,5 @@
  year = {2021},
  publisher = {GitHub},
  journal = {GitHub repository},
- howpublished = {\url {https://github.com/nikita-astronaut/QSL\_at\_QC}},
-}
-
-@misc{amd1999,
- author = {{AMD Corporation}},
- title = {{AMD} Discloses New Technologies at {Microprocessor Forum} --- Including x86 64-bit Architecture and {Lightning Data Transport}$^{{\sc TM}}$},
- howpublished = {Press release},
- day = {5},
- month = oct,
- year = {1999},
- url = {\url{http://www.amd.com/news/prodpr/99105.html}},
+ howpublished = {\url{https://github.com/nikita-astronaut/QSL\_at\_QC}},
 }
