### 2021-08-20 twesterhout/lattice-symmetries#1 @twesterhout
Another update.

* Halide is awesome for 64-bit stuff (i.e. when we're dealing with 64 spins at most). I have implemented kernels for `is_representative` and `state_info`. These are in principle sufficient for all ED stuff (if we replace binary search with the one from STL). The performance is also quite good. On my laptop with avx2 Halide kernels run up to 1.5 times faster than my hand-written ones.
* 512-bit manipulation stuff sucks... I need to be able to bit-shift 512-bit lanes (mind you, not 64-bit lanes, but really all 512 bits together). I failed to make Halide generate reasonable assembly for it, so I guess I'll need to fall back to SIMDe for this. I think I can get a version with 128-bit vectors (i.e. SSE2 and NEON) up and running, but I'll need help with fancier SVE versions since SIMDe doesn't yet support it very well.
