### 2021-08-20 twesterhout/lattice-symmetries#1 @jeffhammond
That's interesting.  @jlinford will have better advice than me on SVE porting.  Is the 512b case addressed by SIMDe support that maps AVX-512 w/ ZMM to SVE512?
