### 2021-08-20 #3537 @jedbrown
@twesterhout Thanks for your revisions. To move forward, could you please:
- [ ] Make a tagged release of your software, and list the version tag of the archived version here. This could be v0.8.2 or v0.9.0 (I see a bug fix and a couple features since v0.8.1). This should be an [annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging).
