### 2021-08-23 #3537 @twesterhout
@jedbrown DOI of the archived version (0.8.2) is 10.5281/zenodo.5235336.

p.s.: I decided to go with 0.8.2 rather than 0.9.0 because the new features are not public yet (they are meant for internal testing / benchmarks).
