### 2021-08-24 bb0b2bd
Add a possibility to extract permutations
