### 2021-09-29 067e187
Expose binary search; Get started with flat_spin_basis for easier interop with Chapel
