+++ src/basis.cpp
@@ -130,6 +130,217 @@ struct ls_states {
     ~ls_states() { ls_destroy_spin_basis(parent); }
 };
 
+struct ls_flat_group {
+    unsigned  shape[3];
+    uint64_t* masks;
+    uint64_t* shifts;
+    double*   eigenvalues_real;
+    double*   eigenvalues_imag;
+    unsigned* sectors;
+    unsigned* periodicities;
+};
+
+struct ls_flat_spin_basis {
+    mutable atomic_count_t refcount;
+    unsigned               number_spins;
+    int                    hamming_weight;
+    int                    spin_inversion;
+    ls_flat_group          group;
+};
+
+namespace lattice_symmetries {
+
+namespace {
+    auto masks_shape(ls_spin_basis const& basis) noexcept -> std::array<unsigned, 3>
+    {
+        struct get_shape_fn_t {
+            auto operator()(small_basis_t const& b) const noexcept
+            {
+                auto network_depth = 0U;
+                if (!b.batched_symmetries.empty()) {
+                    network_depth = b.batched_symmetries.front().network.depth;
+                }
+                else if (b.other_symmetries.has_value()) {
+                    network_depth = b.other_symmetries->network.depth;
+                }
+                auto number_permutations = static_cast<unsigned>(
+                    b.batched_symmetries.size() * batched_small_symmetry_t::batch_size
+                    + b.number_other_symmetries);
+                auto mask_width = 1U;
+                return std::array<unsigned, 3>{{network_depth, number_permutations, mask_width}};
+            }
+
+            auto operator()(big_basis_t const& b) const noexcept
+            {
+                auto network_depth = 0U;
+                if (!b.symmetries.empty()) { network_depth = b.symmetries.front().network.depth; }
+                auto number_permutations = static_cast<unsigned>(b.symmetries.size());
+                auto mask_width          = 8U;
+                return std::array<unsigned, 3>{{network_depth, number_permutations, mask_width}};
+            }
+        };
+        return std::visit(get_shape_fn_t{}, basis.payload);
+    }
+} // namespace
+
+struct init_flat_group_contents_fn_t {
+    ls_flat_group* g;
+
+    auto operator()(small_basis_t const& b) const noexcept -> void
+    {
+        // Initializing masks
+        auto offset = ptrdiff_t{0};
+        for (auto depth = 0U; depth < g->shape[0]; ++depth) {
+            for (auto const& s : b.batched_symmetries) {
+                for (auto i = 0U; i < batched_small_symmetry_t::batch_size; ++i) {
+                    g->masks[offset] = s.network.masks[depth][i];
+                    ++offset;
+                }
+            }
+            for (auto i = 0U; i < b.number_other_symmetries; ++i) {
+                g->masks[offset] = b.other_symmetries->network.masks[depth][i];
+                ++offset;
+            }
+        }
+
+        // Initializing shifts
+        if (g->shifts != nullptr) {
+            auto const& s =
+                !b.batched_symmetries.empty() ? b.batched_symmetries.front() : *b.other_symmetries;
+            for (auto depth = 0U; depth < g->shape[0]; ++depth) {
+                g->shifts[depth] = s.network.deltas[depth];
+            }
+        }
+
+        // Initializing eigenvalues, sectors, and periodicities
+        offset = ptrdiff_t{0};
+        for (auto const& s : b.batched_symmetries) {
+            for (auto i = 0U; i < batched_small_symmetry_t::batch_size; ++i) {
+                g->eigenvalues_real[offset] = s.eigenvalues_real[i];
[…]
