### 2021-09-29 bc3991f
Implement to_csr for Operator
