### 2021-09-30 8544ffd
Flat basis serialization
