### 2021-09-30 afc9947
Python wrappers and tests for FlatSpinBasis
