+++ src/basis.cpp
@@ -153,14 +153,88 @@ struct ls_states {
     ~ls_states() { ls_destroy_spin_basis(parent); }
 };
 
+namespace lattice_symmetries {
+
+template <bool CallDestructor = true> struct free_deleter_fn_t {
+    template <class T> auto operator()(T* ptr) const noexcept -> void
+    {
+        if (ptr != nullptr) {
+            ptr->~T();
+            std::free(ptr);
+        }
+    }
+};
+
+template <class T>
+auto alloc_aligned_array(size_t alignment, size_t count) noexcept
+    -> std::unique_ptr<T, free_deleter_fn_t<false>>
+{
+    auto* ptr = count > 0 ? std::aligned_alloc(alignment, count * sizeof(T)) : nullptr;
+    return std::unique_ptr<T, free_deleter_fn_t<false>>{static_cast<T*>(ptr)};
+}
+
+template <class T, class... Args>
+auto alloc_aligned(Args&&... args) noexcept -> std::unique_ptr<T, free_deleter_fn_t<true>>
+{
+    static_assert(std::is_nothrow_constructible_v<T, Args&&...>);
+    constexpr auto alignment = 64U;
+    auto*          ptr       = std::aligned_alloc(alignment, sizeof(T));
+    new (ptr) T{std::forward<Args>(args)...};
+    return std::unique_ptr<T, free_deleter_fn_t<true>>{static_cast<T*>(ptr)};
+}
+} // namespace lattice_symmetries
+
+struct ls_flat_spin_basis;
+
 struct ls_flat_group {
-    unsigned  shape[3];
-    uint64_t* masks;
-    uint64_t* shifts;
-    double*   eigenvalues_real;
-    double*   eigenvalues_imag;
-    unsigned* sectors;
-    unsigned* periodicities;
+    template <class T>
+    using buffer_t = std::unique_ptr<T, lattice_symmetries::free_deleter_fn_t<false>>;
+    static constexpr uint64_t alignment = 64U;
+    friend struct ls_flat_spin_basis;
+
+    std::array<unsigned, 3> shape;
+    buffer_t<uint64_t>      masks;
+    buffer_t<uint64_t>      shifts;
+    buffer_t<double>        eigenvalues_real;
+    buffer_t<double>        eigenvalues_imag;
+    buffer_t<unsigned>      sectors;
+    buffer_t<unsigned>      periodicities;
+
+  private:
+    explicit ls_flat_group(std::array<unsigned, 3> _shape) noexcept
+        : shape{_shape}
+        , masks{lattice_symmetries::alloc_aligned_array<uint64_t>(alignment,
+                                                                  shape[0] * shape[1] * shape[2])}
+        , shifts{lattice_symmetries::alloc_aligned_array<uint64_t>(alignment, shape[0])}
+        , eigenvalues_real{lattice_symmetries::alloc_aligned_array<double>(alignment, shape[1])}
+        , eigenvalues_imag{lattice_symmetries::alloc_aligned_array<double>(alignment, shape[1])}
+        , sectors{lattice_symmetries::alloc_aligned_array<unsigned>(alignment, shape[1])}
+        , periodicities{lattice_symmetries::alloc_aligned_array<unsigned>(alignment, shape[1])}
+    {}
+
+    [[nodiscard]] auto has_allocation_succeeded() const noexcept -> bool
+    {
+        if (shape[0] * shape[1] * shape[2] != 0 && masks == nullptr) {
+            LATTICE_SYMMETRIES_LOG_DEBUG("Failed to allocate 'masks' array of shape [%u, %u, %u]\n",
+                                         shape[0], shape[1], shape[2]);
+            return false;
+        }
[…]
