### 2021-09-30 f0a5e66
Add Halide kernels for state_info and is_representative (they are not yet integrated)
