+++ src/halide/generator.cpp
@@ -1,0 +1,256 @@
+#include "Halide.h"
+
+using namespace Halide;
+
+auto bit_permute_step_64(Expr x, Expr m, Expr d) -> Expr
+{
+    auto y = ((x >> d) ^ x) & m;
+    return (x ^ y) ^ (y << d);
+}
+
+class state_info_generator : public Halide::Generator<state_info_generator> {
+  public:
+    GeneratorParam<int> _spin_inversion{"spin_inversion", /*default value*/ 0};
+
+    Input<Buffer<uint64_t>>  _x{"x", 1};
+    Input<uint64_t>          _flip_mask{"flip_mask"};
+    Input<Buffer<uint64_t>>  _masks{"masks", 2};
+    Input<Buffer<double>>    _eigvals_re{"eigvals_re", 1};
+    Input<Buffer<double>>    _eigvals_im{"eigvals_im", 1};
+    Input<Buffer<unsigned>>  _shifts{"shifts", 1};
+    Output<Buffer<uint64_t>> _repr{"representative", 1};
+    Output<Buffer<double>>   _character{"character", 2};
+    Output<Buffer<double>>   _norm{"norm"};
+
+    auto reduction_step_impl(Expr i, Tuple current, Expr y, Expr c_re, Expr c_im) const -> Tuple
+    {
+        auto current_r    = current[0];
+        auto current_c_re = current[1];
+        auto current_c_im = current[2];
+        auto current_n    = current[3];
+
+        auto is_less   = y < current_r;
+        auto next_r    = select(is_less, y, current_r);
+        auto next_c_re = select(is_less, c_re, current_c_re);
+        auto next_c_im = select(is_less, c_im, current_c_im);
+
+        auto is_equal = y == _x(i);
+        auto next_n   = select(is_equal, current_n + c_re, current_n);
+
+        return {next_r, next_c_re, next_c_im, next_n};
+    }
+
+    auto reduction_step(Expr i, Tuple current, Expr y, Expr c_re, Expr c_im) const -> Tuple
+    {
+        if (_spin_inversion == 0) { return reduction_step_impl(i, current, y, c_re, c_im); }
+        if (_spin_inversion == 1) {
+            auto temp = reduction_step_impl(i, current, y, c_re, c_im);
+            return reduction_step_impl(i, temp, y ^ _flip_mask, c_re, c_im);
+        }
+        if (_spin_inversion == -1) {
+            auto temp = reduction_step_impl(i, current, y, c_re, c_im);
+            return reduction_step_impl(i, temp, y ^ _flip_mask, -c_re, -c_im);
+        }
+        throw std::runtime_error{"invalid spin_inversion"};
+    }
+
+    auto reduction_step(Tuple current, Tuple other) const -> Tuple
+    {
+        auto current_r    = current[0];
+        auto current_c_re = current[1];
+        auto current_c_im = current[2];
+        auto current_n    = current[3];
+        auto r            = other[0];
+        auto c_re         = other[1];
+        auto c_im         = other[2];
+        auto n            = other[3];
+
+        auto is_less   = r < current_r;
+        auto next_r    = select(is_less, r, current_r);
+        auto next_c_re = select(is_less, c_re, current_c_re);
+        auto next_c_im = select(is_less, c_im, current_c_im);
+        auto next_n    = current_n + n;
+
+        return {next_r, next_c_re, next_c_im, next_n};
+    }
+
+    void generate()
+    {
+        auto       depth         = _masks.dim(0).extent();
+        auto       number_masks  = _masks.dim(1).extent();
+        auto const chunk_size    = 2 * natural_vector_size(type_of<uint64_t>());
[…]
