### 2021-09-30 73edfe1
Enable arch-specific kernels in Halide
