### 2021-10-02 4980afc
Export is_representative; try fixing compilation on OS X
