### 2021-10-11 0f6f563
Make C++ code compatible with Halide kernels
