### 2021-10-17 9a0e9aa
is_representative kernel was missing final sclar reduction
