### 2021-10-22 391ae97
loading SpinBasis from YAML
