+++ src/LatticeSymmetries.hs
@@ -1,9 +1,25 @@
-module LatticeSymmetries
-  ( foo,
-  )
-where
+{-# LANGUAGE CApiFFI #-}
 
-import Foreign.C.Types (CInt (..))
+module LatticeSymmetries where
+
+import Data.Aeson
+import Data.Aeson.Types (typeMismatch)
+import Data.Complex
+import qualified Data.List.NonEmpty as NonEmpty
+import Data.Scientific (toRealFloat)
+import qualified Data.Vector.Unboxed as U
+import Data.Yaml (decodeFileWithWarnings)
+import Foreign.C.String (CString, peekCString)
+import Foreign.C.Types (CInt (..), CUInt (..))
+import Foreign.ForeignPtr
+import Foreign.Marshal.Alloc (alloca)
+import Foreign.Marshal.Array (withArrayLen)
+import Foreign.Ptr (FunPtr, Ptr)
+import Foreign.Storable (Storable (..))
+import qualified GHC.ForeignPtr as GHC (Finalizers (..), ForeignPtr (..), ForeignPtrContents (..))
+import qualified GHC.IORef as GHC (atomicSwapIORef)
+import qualified GHC.Ptr as GHC (Ptr (..))
+import UnliftIO.Exception (bracket, impureThrow, throwIO)
 
 foo :: IO CInt
 foo = do
@@ -11,3 +27,290 @@ foo = do
   return 123
 
 foreign export ccall "ls_hs_foo" foo :: IO CInt
+
+-- | Exceptions thrown when an error occurs in @liblattice_symmetries@.
+data LatticeSymmetriesException = LatticeSymmetriesException {eCode :: Int, eMessage :: Text}
+  deriving stock (Show)
+
+instance Exception LatticeSymmetriesException
+
+data SpinEDException = SpinEDException Text
+  deriving stock (Show)
+
+instance Exception SpinEDException
+
+-- | Retrieve textual representation of an error
+getErrorMessage ::
+  -- | Error code returned by @lattice_symmetries@ library
+  Int ->
+  -- | Explanation of the error
+  IO Text
+getErrorMessage c = bracket (ls_error_to_string (fromIntegral c)) ls_destroy_string $ \s ->
+  toText <$> peekCString s
+
+-- | Check the status code returned by @lattice_symmetries@ library. If it
+-- indicates an error, 'LatticeSymmetriesException' is thrown.
+checkStatus :: (MonadIO m, Integral a) => a -> m ()
+checkStatus c
+  | c == 0 = return ()
+  | otherwise = do
+    print =<< liftIO (getErrorMessage c')
+    throwIO . LatticeSymmetriesException c' =<< liftIO (getErrorMessage c')
+  where
+    c' = fromIntegral c
+
+data SymmetrySpec = SymmetrySpec !(NonEmpty Int) !Int
+  deriving stock (Read, Show, Eq)
+
+instance FromJSON SymmetrySpec where
+  parseJSON = withObject "symmetry" $ \v ->
+    SymmetrySpec
+      <$> v .: "permutation"
+      <*> v .: "sector"
+
+instance ToJSON SymmetrySpec where
+  toJSON (SymmetrySpec permutation sector) =
+    object ["permutation" .= permutation, "sector" .= sector]
+
+toSymmetry :: MonadIO m => SymmetrySpec -> m Symmetry
+toSymmetry (SymmetrySpec p s) = mkSymmetry p s
+
+data BasisSpec = BasisSpec !Int !(Maybe Int) !(Maybe Int) ![SymmetrySpec]
+  deriving stock (Read, Show, Eq)
+
[…]
