+++ cbits/lattice_symmetries_haskell.h
@@ -6,6 +6,8 @@
 #include <stdint.h>
 #endif
 
+#include <lattice_symmetries/lattice_symmetries.h>
+
 #if defined(__cplusplus)
 extern "C" {
 #endif
@@ -13,6 +15,7 @@ extern "C" {
 void ls_hs_init(void);
 void ls_hs_exit(void);
 int ls_hs_foo(void);
+ls_spin_basis *ls_hs_load_basis_from_yaml(char const *path);
 
 #ifdef __cplusplus
 } // extern "C"
