+++ test/Spec.hs
@@ -1,0 +1,190 @@
+module Main (main) where
+
+import Data.Complex
+import qualified Data.Vector.Storable as S
+import Data.Yaml.Aeson
+import Foreign.Storable
+import LatticeSymmetries
+import Test.Hspec
+
+anySpinEDException :: Selector SpinEDException
+anySpinEDException = const True
+
+anyLatticeSymmetriesException :: Selector LatticeSymmetriesException
+anyLatticeSymmetriesException = const True
+
+parseLines :: FromJSON a => [Text] -> Either ParseException a
+parseLines xs = decodeEither' . encodeUtf8 . unlines $ xs
+
+checkRight s y = case s of
+  Left _ -> s `shouldSatisfy` isRight
+  Right x -> x `shouldBe` y
+
+checkJust s y = case s of
+  Just x -> x `shouldBe` y
+  Nothing -> s `shouldSatisfy` isJust
+
+main :: IO ()
+main = hspec $ do
+  describe "SymmetrySpec" $ do
+    it "parses JSON specifications" $ do
+      let s₁ =
+            parseLines @SymmetrySpec
+              [ "permutation: [3, 1, 2, 4]",
+                "sector: 8"
+              ]
+          s₂ =
+            parseLines @SymmetrySpec
+              [ "permutation: [1, 1, 2, 4]",
+                "sector: -5"
+              ]
+          s₃ =
+            parseLines @SymmetrySpec
+              [ "permutation: []",
+                "sector: 5"
+              ]
+          s₄ =
+            parseLines @SymmetrySpec
+              [ "permutation: [1]",
+                "sector: null"
+              ]
+      case s₁ of
+        Left _ -> s₁ `shouldSatisfy` isRight
+        Right x -> x `shouldBe` SymmetrySpec (fromList [3, 1, 2, 4]) 8
+      case s₂ of
+        Left _ -> s₂ `shouldSatisfy` isRight
+        Right x -> x `shouldBe` SymmetrySpec (fromList [1, 1, 2, 4]) (-5)
+      s₃ `shouldSatisfy` isLeft
+      s₄ `shouldSatisfy` isLeft
+  describe "denseMatrixFromList" $ do
+    it "handles empty lists" $ do
+      denseMatrixFromList [] `shouldBe` Just (DenseMatrix (0, 0) S.empty)
+      denseMatrixFromList [[], []] `shouldBe` Just (DenseMatrix (2, 0) S.empty)
+    it "handles square matrices" $ do
+      denseMatrixFromList [[1, 2], [3, 4]] `shouldBe` Just (DenseMatrix (2, 2) (fromList [1, 2, 3, 4]))
+    it "handles rectangular matrices" $ do
+      denseMatrixFromList [[1, 2, 3], [4, 5, 6]] `shouldBe` Just (DenseMatrix (2, 3) (fromList [1, 2, 3, 4, 5, 6]))
+      denseMatrixFromList [[1, 2], [3, 4], [5, 6]] `shouldBe` Just (DenseMatrix (3, 2) (fromList [1, 2, 3, 4, 5, 6]))
+    it "general lists" $ do
+      denseMatrixFromList [[1, 2], [3], [5, 6]] `shouldBe` Nothing
+  describe "DenseMatrixSpec" $ do
+    it "parses JSON specifications" $ do
+      parseLines @DenseMatrixSpec
+        [ "[[1, 2,  1],",
+          " [4, 2,  1],",
+          " [1, -2, 3]]"
+        ]
+        `checkRight` DenseMatrixSpec [[1, 2, 1], [4, 2, 1], [1, -2, 3]]
+      parseLines @DenseMatrixSpec
[…]
