### 2021-10-23 3ea67d9
Started implementing sparse operator
