+++ src/LatticeSymmetries.hs
@@ -1,12 +1,22 @@
 {-# LANGUAGE CApiFFI #-}
+{-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+-- {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module LatticeSymmetries where
 
+import qualified Control.Monad.Primitive as Primitive
+import Control.Monad.ST
+import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
 import Data.Aeson
 import Data.Aeson.Types (typeMismatch)
 import Data.Complex
 import qualified Data.List.NonEmpty as NonEmpty
 import Data.Scientific (toRealFloat)
+import Data.Vector.Generic ((!))
+import qualified Data.Vector.Storable as S
+import qualified Data.Vector.Storable.Mutable as SM
 import qualified Data.Vector.Unboxed as U
 import Data.Yaml (decodeFileWithWarnings)
 import Foreign.C.String (CString, peekCString)
@@ -16,10 +26,19 @@ import Foreign.Marshal.Alloc (alloca)
 import Foreign.Marshal.Array (withArrayLen)
 import Foreign.Ptr (FunPtr, Ptr)
 import Foreign.Storable (Storable (..))
+import qualified GHC.Exts as GHC (IsList (..))
 import qualified GHC.ForeignPtr as GHC (Finalizers (..), ForeignPtr (..), ForeignPtrContents (..))
 import qualified GHC.IORef as GHC (atomicSwapIORef)
 import qualified GHC.Ptr as GHC (Ptr (..))
+import qualified Language.C.Inline as C
+import qualified Language.C.Inline.Unsafe as CU
+import LatticeSymmetries.Context
+import qualified System.IO.Unsafe
 import UnliftIO.Exception (bracket, impureThrow, throwIO)
+
+C.context (C.baseCtx <> C.bsCtx <> C.funCtx <> lsCtx)
+C.include "<lattice_symmetries/lattice_symmetries.h>"
+C.include "helpers.h"
 
 foo :: IO CInt
 foo = do
@@ -110,6 +129,59 @@ instance FromJSON (Complex Double) where
     _ -> typeMismatch "Complex" v
   parseJSON v = typeMismatch "Complex" v
 
+newtype DenseMatrixSpec = DenseMatrixSpec [[Complex Double]]
+  deriving stock (Read, Show, Eq)
+  deriving newtype (FromJSON)
+
+-- Row-major order (C layout)
+data DenseMatrix = DenseMatrix {denseMatrixShape :: (Int, Int), denseMatrixData :: S.Vector (Complex Double)}
+  deriving stock (Show, Eq)
+
+isDenseMatrixSquare :: DenseMatrix -> Bool
+isDenseMatrixSquare (DenseMatrix (r, c) _) = r == c
+
+isDenseMatrixEmpty :: DenseMatrix -> Bool
+isDenseMatrixEmpty (DenseMatrix (r, c) _) = r * c == 0
+
+indexDenseMatrix :: DenseMatrix -> Int -> Int -> Complex Double
+indexDenseMatrix (DenseMatrix (_, c) v) i j = v ! (c * i + j)
+
+isDenseMatrixHermitian :: DenseMatrix -> Bool
+isDenseMatrixHermitian matrix@(DenseMatrix (r, c) _) = isDenseMatrixSquare matrix && go 0 0
+  where
+    -- Iterate over upper triangle (including the diagonal) of the matrix
+    go :: Int -> Int -> Bool
+    go !i !j
+      | j == c = let !i' = i + 1 in (i' >= r) || go i' i'
+      | otherwise =
[…]
