### 2021-10-23 654cae1
inline-c context; some ls_sparse_operator & ls_workspace types
