### 2021-10-25 b6793fa
Implement is_representative and state_info kernels for flat_spin_basis without symmetries
