### 2021-10-26 ab8c7dd
Now it's a simple copy-paste from SpinED
