+++ src/LatticeSymmetries.hs
@@ -1,62 +1,24 @@
 {-# LANGUAGE CApiFFI #-}
-{-# LANGUAGE QuasiQuotes #-}
-{-# LANGUAGE TemplateHaskell #-}
 
--- {-# OPTIONS_GHC -fno-warn-orphans #-}
+-- {-# LANGUAGE QuasiQuotes #-}
+-- {-# LANGUAGE TemplateHaskell #-}
 
 module LatticeSymmetries where
 
-import qualified Control.Monad.Primitive as Primitive
-import Control.Monad.ST
-import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
-import Data.Aeson
-import Data.Aeson.Types (typeMismatch)
 import Data.Complex
-import qualified Data.List.NonEmpty as NonEmpty
-import Data.Scientific (toRealFloat)
-import Data.Vector.Generic ((!))
-import qualified Data.Vector.Storable as S
-import qualified Data.Vector.Storable.Mutable as SM
-import qualified Data.Vector.Unboxed as U
-import Data.Yaml (decodeFileWithWarnings)
+import Data.Vector.Storable (Vector)
+import qualified Data.Vector.Storable as V
 import Foreign.C.String (CString, peekCString)
-import Foreign.C.Types (CInt (..), CUInt (..))
+import Foreign.C.Types (CBool (..), CInt (..), CUInt (..), CUShort (..))
 import Foreign.ForeignPtr
 import Foreign.Marshal.Alloc (alloca)
 import Foreign.Marshal.Array (withArrayLen)
-import Foreign.Ptr (FunPtr, Ptr)
+import Foreign.Ptr (FunPtr, Ptr, nullPtr)
 import Foreign.Storable (Storable (..))
-import qualified GHC.Exts as GHC (IsList (..))
-import qualified GHC.ForeignPtr as GHC (Finalizers (..), ForeignPtr (..), ForeignPtrContents (..))
-import qualified GHC.IORef as GHC (atomicSwapIORef)
-import qualified GHC.Ptr as GHC (Ptr (..))
-import qualified Language.C.Inline as C
-import qualified Language.C.Inline.Unsafe as CU
-import LatticeSymmetries.Context
+import LatticeSymmetries.IO
+import LatticeSymmetries.Types
 import qualified System.IO.Unsafe
 import UnliftIO.Exception (bracket, impureThrow, throwIO)
-
-C.context (C.baseCtx <> C.bsCtx <> C.funCtx <> lsCtx)
-C.include "<lattice_symmetries/lattice_symmetries.h>"
-C.include "helpers.h"
-
-foo :: IO CInt
-foo = do
-  putStrLn "Hello world!"
-  return 123
-
-foreign export ccall "ls_hs_foo" foo :: IO CInt
-
--- | Exceptions thrown when an error occurs in @liblattice_symmetries@.
-data LatticeSymmetriesException = LatticeSymmetriesException {eCode :: Int, eMessage :: Text}
-  deriving stock (Show)
-
-instance Exception LatticeSymmetriesException
-
-data SpinEDException = SpinEDException Text
-  deriving stock (Show)
-
-instance Exception SpinEDException
 
 -- | Retrieve textual representation of an error
 getErrorMessage ::
@@ -78,124 +40,20 @@ checkStatus c
   where
     c' = fromIntegral c
 
-data SymmetrySpec = SymmetrySpec !(NonEmpty Int) !Int
-  deriving stock (Read, Show, Eq)
+-- loadRawBasis :: MonadIO m => Text -> m (Ptr Cspin_basis)
+-- loadRawBasis path = do
+--   r <- liftIO $ decodeFileWithWarnings (toString path)
+--   case r of
[…]
