### 2021-10-26 50c8055
Move all new stuff to LatticeSymmetries.Sparse
