+++ cbits/lattice_symmetries_haskell.h
@@ -14,8 +14,58 @@ extern "C" {
 
 void ls_hs_init(void);
 void ls_hs_exit(void);
-int ls_hs_foo(void);
-ls_spin_basis *ls_hs_load_basis_from_yaml(char const *path);
+int ls_hs_foo(int);
+
+void *ls_hs_internal_load_basis_from_yaml(void *);
+inline ls_spin_basis *ls_hs_load_basis_from_yaml(char const *path) {
+  return ls_hs_internal_load_basis_from_yaml((void *)path);
+}
+
+void *ls_hs_get_lowered_operator(void *);
+
+typedef struct ls_sparse_operator ls_sparse_operator;
+typedef struct ls_term ls_term;
+
+typedef struct ls_hs_basis {
+  ls_flat_spin_basis const *payload;
+  void *context;
+} ls_hs_basis;
+
+typedef struct ls_hs_operator {
+  ls_sparse_operator const *payload;
+  void *context;
+} ls_hs_operator;
+
+typedef void (*ls_spin_copy_fn)(uint64_t const * /*source*/,
+                                uint64_t * /*destination*/);
+typedef void (*ls_spin_fill_fn)(uint64_t const * /*2D source*/,
+                                uint64_t /*count*/,
+                                uint64_t * /*2D destination*/);
+
+typedef struct ls_output_buffer {
+  uint64_t *spins;
+  _Complex double *coeffs;
+  _Complex double *const diagonal;
+  uint64_t const number_words;
+  ls_spin_copy_fn const spin_copy;
+  ls_spin_fill_fn const spin_fill;
+} ls_output_buffer;
+
+typedef struct ls_workspace {
+  uint64_t *spins;
+  _Complex double *characters;
+  double *norms;
+} ls_workspace;
+
+void ls_hs_create_operator(ls_hs_operator *self,
+                           ls_flat_spin_basis const *basis,
+                           unsigned number_terms, ls_term const *term);
+void ls_hs_destroy_operator(ls_hs_operator *self);
+void ls_hs_load_hamiltonian_from_yaml(ls_hs_operator *self, char const *path);
+void ls_hs_operator_get_basis(ls_hs_operator const *self, ls_hs_basis *basis);
+
+uint64_t ls_hs_apply_operator(ls_hs_operator const *term, uint64_t const *spin,
+                              ls_output_buffer *out, ls_workspace *workspace);
 
 #ifdef __cplusplus
 } // extern "C"
