+++ src/LatticeSymmetries/Sparse.hs
@@ -1,0 +1,640 @@
+{-# LANGUAGE CApiFFI #-}
+{-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+module LatticeSymmetries.Sparse where
+
+import qualified Control.Monad.Primitive as Primitive
+import Control.Monad.ST
+import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
+import Data.Aeson
+import Data.Aeson.Types (typeMismatch)
+import Data.Binary (Binary (..))
+import Data.Complex
+import qualified Data.List.NonEmpty as NonEmpty
+import Data.Scientific (toRealFloat)
+import Data.Vector.Binary
+import Data.Vector.Generic ((!))
+import qualified Data.Vector.Storable as S
+import qualified Data.Vector.Storable.Mutable as SM
+import qualified Data.Vector.Unboxed as U
+import Data.Yaml (decodeFileWithWarnings)
+import Foreign.C.String (CString, peekCString)
+import Foreign.C.Types (CInt (..), CUInt (..))
+import Foreign.ForeignPtr
+import Foreign.ForeignPtr.Unsafe (unsafeForeignPtrToPtr)
+import Foreign.Marshal.Alloc (alloca, free, malloc, mallocBytes)
+import Foreign.Marshal.Array (newArray, withArrayLen)
+import Foreign.Marshal.Utils (new, with)
+import Foreign.Ptr (FunPtr, Ptr, castPtr)
+import Foreign.StablePtr
+import Foreign.Storable (Storable (..))
+import qualified GHC.Exts as GHC (IsList (..))
+import qualified GHC.ForeignPtr as GHC (Finalizers (..), ForeignPtr (..), ForeignPtrContents (..))
+import GHC.Generics
+import qualified GHC.IORef as GHC (atomicSwapIORef)
+import GHC.Prim
+import qualified GHC.Ptr as GHC (Ptr (..))
+import qualified Language.C.Inline as C
+import qualified Language.C.Inline.Unsafe as CU
+import LatticeSymmetries.Context
+import LatticeSymmetries.IO
+import LatticeSymmetries.Types
+import qualified System.IO.Unsafe
+import qualified System.Mem.Weak
+import UnliftIO.Exception (bracket, impureThrow, throwIO)
+
+C.context (C.baseCtx <> C.bsCtx <> C.funCtx <> lsCtx)
+C.include "<lattice_symmetries/lattice_symmetries.h>"
+C.include "helpers.h"
+
+-- Row-major order (C layout)
+data DenseMatrix a = DenseMatrix {denseMatrixShape :: (Int, Int), denseMatrixData :: S.Vector a}
+  deriving stock (Show, Eq, Generic)
+  deriving anyclass (Binary)
+
+isDenseMatrixSquare :: DenseMatrix a -> Bool
+isDenseMatrixSquare (DenseMatrix (r, c) _) = r == c
+
+isDenseMatrixEmpty :: DenseMatrix a -> Bool
+isDenseMatrixEmpty (DenseMatrix (r, c) _) = r * c == 0
+
+indexDenseMatrix :: Storable a => DenseMatrix a -> Int -> Int -> a
+indexDenseMatrix (DenseMatrix (_, c) v) i j = v ! (c * i + j)
+
+isDenseMatrixHermitian :: (Storable a, Num a, Eq a) => DenseMatrix (Complex a) -> Bool
+isDenseMatrixHermitian matrix@(DenseMatrix (r, c) _) = isDenseMatrixSquare matrix && go 0 0
+  where
[…]
