### 2021-10-26 7ff64a5
Update build scripts
