### 2021-10-27 c339096
Compile init.c with -fPIC, add more debugging output
