### 2021-12-02 a8f44c8
Add simple operations with hdf5 files
