+++ cbits/lattice_symmetries_haskell.h
@@ -31,6 +31,12 @@ void ls_hs_basis_and_hamiltonian_from_yaml(char const *path,
 void ls_hs_destroy_spin_basis(ls_hs_spin_basis_v1 *basis);
 void ls_hs_destroy_operator(ls_hs_operator_v1 *op);
 
+void ls_hs_hdf5_create_dataset_u64(char const *filename, char const *dataset,
+                                   unsigned dim, uint64_t const *shape);
+void ls_hs_hdf5_write_1d_chunk_u64(char const *filename, char const *dataset,
+                                   uint64_t offset, uint64_t count,
+                                   uint64_t const *data);
+
 #if 0
 typedef struct ls_sparse_operator ls_sparse_operator;
 typedef struct ls_term ls_term;
