+++ src/LatticeSymmetries.hs
@@ -14,7 +14,7 @@ import Foreign.C.String (CString, peekCString)
 import Foreign.C.Types (CBool (..), CInt (..), CUInt (..), CUShort (..))
 import Foreign.ForeignPtr
 import Foreign.Marshal.Alloc (alloca)
-import Foreign.Marshal.Array (peekArray, withArrayLen)
+import Foreign.Marshal.Array (peekArray, pokeArray, withArrayLen)
 import Foreign.Ptr (FunPtr, Ptr, nullPtr)
 import Foreign.StablePtr
 import Foreign.Storable (Storable (..))
@@ -95,39 +95,158 @@ createDataset _fileName _datasetName _dim shapePtr = do
 ls_hs_hdf5_create_dataset_u64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
 ls_hs_hdf5_create_dataset_u64 = createDataset @Word64
 
+ls_hs_hdf5_create_dataset_f32 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+ls_hs_hdf5_create_dataset_f32 = createDataset @Double
+
+ls_hs_hdf5_create_dataset_f64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+ls_hs_hdf5_create_dataset_f64 = createDataset @Double
+
+ls_hs_hdf5_create_dataset_c64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+ls_hs_hdf5_create_dataset_c64 = createDataset @(Complex Float)
+
+ls_hs_hdf5_create_dataset_c128 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+ls_hs_hdf5_create_dataset_c128 = createDataset @(Complex Double)
+
 foreign export ccall "ls_hs_hdf5_create_dataset_u64"
   ls_hs_hdf5_create_dataset_u64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+
+foreign export ccall "ls_hs_hdf5_create_dataset_f32"
+  ls_hs_hdf5_create_dataset_f32 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+
+foreign export ccall "ls_hs_hdf5_create_dataset_f64"
+  ls_hs_hdf5_create_dataset_f64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+
+foreign export ccall "ls_hs_hdf5_create_dataset_c64"
+  ls_hs_hdf5_create_dataset_c64 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
+
+foreign export ccall "ls_hs_hdf5_create_dataset_c128"
+  ls_hs_hdf5_create_dataset_c128 :: CString -> CString -> CUInt -> Ptr Word64 -> IO ()
 
 writeDatasetChunk ::
   forall a.
   (Storable a, H5.KnownDatatype a) =>
   CString ->
   CString ->
-  Word64 ->
-  Word64 ->
+  CUInt ->
+  Ptr Word64 ->
+  Ptr Word64 ->
   Ptr a ->
   IO ()
-writeDatasetChunk _fileName _datasetName _offset _size ptr = do
+writeDatasetChunk _fileName _datasetName _dim _offsetPtr _shapePtr dataPtr = do
   fileName <- fromString <$> peekCString _fileName
   datasetName <- fromString <$> peekCString _datasetName
-  -- shape <- fmap fromIntegral <$> peekArray (fromIntegral _dim) shapePtr
-  fp <- newForeignPtr_ ptr
[…]
