### 2021-12-14 a2fa4eb
Link against lattice_symmetries
