### 2021-12-14 fcc966a
implement ls_hs_operator_apply and ls_hs_basis_index
