### 2021-12-16 692aeb8
Link hdf5 statically
