### 2021-12-17 b42b894
Bump Halide version
