### 2021-12-17 5dfd08e
Continue work on Sparse module
