+++ src/LatticeSymmetries/Sparse.hs
@@ -4,17 +4,23 @@
 
 module LatticeSymmetries.Sparse where
 
+-- import Data.Binary (Binary (..))
+
+-- import Data.Vector.Binary
+
+import Control.Exception.Safe (bracket, impureThrow, throwIO)
 import qualified Control.Monad.Primitive as Primitive
 import Control.Monad.ST
 import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
 import Data.Aeson
 import Data.Aeson.Types (typeMismatch)
-import Data.Binary (Binary (..))
 import Data.Complex
+import qualified Data.List
 import qualified Data.List.NonEmpty as NonEmpty
 import Data.Scientific (toRealFloat)
-import Data.Vector.Binary
+import qualified Data.Vector as B
 import Data.Vector.Generic ((!))
+import qualified Data.Vector.Generic as G
 import qualified Data.Vector.Storable as S
 import qualified Data.Vector.Storable.Mutable as SM
 import qualified Data.Vector.Unboxed as U
@@ -42,7 +48,6 @@ import LatticeSymmetries.IO
 import LatticeSymmetries.Types
 import qualified System.IO.Unsafe
 import qualified System.Mem.Weak
-import UnliftIO.Exception (bracket, impureThrow, throwIO)
 
 C.context (C.baseCtx <> C.bsCtx <> C.funCtx <> lsCtx)
 C.include "<lattice_symmetries/lattice_symmetries.h>"
@@ -51,7 +56,135 @@ C.include "helpers.h"
 -- Row-major order (C layout)
 data DenseMatrix a = DenseMatrix {denseMatrixShape :: (Int, Int), denseMatrixData :: S.Vector a}
   deriving stock (Show, Eq, Generic)
-  deriving anyclass (Binary)
+
+-- deriving anyclass (Binary)
+
+data SparseMatrix a = SparseMatrix
+  { smRows :: !Int,
+    smCols :: !Int,
+    smOffsets :: !(S.Vector CUInt),
+    smIndices :: !(S.Vector CUInt),
+    smData :: !(S.Vector a)
+  }
+  deriving stock (Show, Eq, Generic)
+
+binarySearch :: (G.Vector v a, Ord a) => v a -> a -> Maybe Int
+binarySearch v z = go 0 (G.length v)
+  where
+    go !l !u
+      | l < u =
+        -- NOTE: we assume that the vector is short enought such that @u + l@ does not overflow
+        let !m = (u + l) `div` 2
+         in case compare (v ! m) z of
+              LT -> go (m + 1) u
+              EQ -> Just m
+              GT -> go l m
+      | otherwise = Nothing
+
+smIndex :: (Num a, Storable a) => SparseMatrix a -> (Int, Int) -> a
+smIndex matrix (i, j) =
+  case binarySearch (G.slice l (u - l) (smIndices matrix)) (fromIntegral j) of
+    Just k -> smData matrix ! (l + k)
+    Nothing -> 0
+  where
+    !l = fromIntegral $ smOffsets matrix ! i
+    !u = fromIntegral $ smOffsets matrix ! (i + 1)
+{-# INLINE smIndex #-}
+
+preprocessCoo :: Num a => [(Int, Int, a)] -> [(Int, Int, a)]
+preprocessCoo =
+  fmap (Data.List.foldl1' (\(!i, !j, !x) (_, _, y) -> (i, j, x + y)))
+    . Data.List.groupBy (\a b -> key a == key b)
+    . sortOn key
+  where
+    key (i, j, _) = (i, j)
+
+unsafeCooToCsr :: forall a. (Storable a, Num a) => [(Int, Int, a)] -> SparseMatrix a
[…]
