### 2021-12-19 c61feba
Playing around with operator strings
