+++ test/Spec.hs
@@ -2,6 +2,7 @@ module Main (main) where
 
 import Data.Complex
 import qualified Data.Vector.Storable as S
+import qualified Data.Vector.Unboxed as U
 import Data.Yaml.Aeson
 import Foreign.Storable
 import LatticeSymmetries
@@ -63,19 +64,18 @@ main = hspec $ do
       s₄ `shouldSatisfy` isLeft
   describe "denseMatrixFromList" $ do
     it "handles empty lists" $ do
-      denseMatrixFromList [] `shouldBe` Right (DenseMatrix (0, 0) S.empty :: DenseMatrix Double)
-      denseMatrixFromList [[], []] `shouldBe` Right (DenseMatrix (2, 0) S.empty :: DenseMatrix Double)
+      denseMatrixFromList []
+        `shouldBe` Right (DenseMatrix S.empty :: DenseMatrix S.Vector 0 0 Double)
+      denseMatrixFromList [[], []]
+        `shouldBe` Right (DenseMatrix S.empty :: DenseMatrix S.Vector 2 0 Double)
     it "handles square matrices" $ do
       denseMatrixFromList [[1, 2], [3, 4]]
-        `shouldBe` Right (DenseMatrix (2, 2) (fromList [1, 2, 3, 4]) :: DenseMatrix Double)
+        `shouldBe` Right (DenseMatrix (fromList [1, 2, 3, 4]) :: DenseMatrix S.Vector 2 2 Double)
     it "handles rectangular matrices" $ do
       denseMatrixFromList [[1, 2, 3], [4, 5, 6]]
-        `shouldBe` Right (DenseMatrix (2, 3) (fromList [1, 2, 3, 4, 5, 6]) :: DenseMatrix Double)
+        `shouldBe` Right (DenseMatrix (fromList [1, 2, 3, 4, 5, 6]) :: DenseMatrix S.Vector 2 3 Double)
       denseMatrixFromList [[1, 2], [3, 4], [5, 6]]
-        `shouldBe` Right (DenseMatrix (3, 2) (fromList [1, 2, 3, 4, 5, 6]) :: DenseMatrix Double)
-  -- it "general lists" $ do
-  --   (denseMatrixFromList [[1, 2], [3], [5, 6]] :: Maybe (DenseMatrix Int))
-  --     `shouldBe` Nothing
+        `shouldBe` Right (DenseMatrix (fromList [1, 2, 3, 4, 5, 6]) :: DenseMatrix S.Vector 3 2 Double)
   describe "DenseMatrixSpec" $ do
     it "parses JSON specifications" $ do
       parseLines @DenseMatrixSpec
@@ -157,32 +157,85 @@ main = hspec $ do
       binarySearch (S.fromList [1]) (1 :: Int) `shouldBe` Just 0
   describe "cooToCSR" $ do
     it ".." $ do
-      let m = cooToCsr [(2, 2, -3), (1, 1, 1), (2, 3, 8), (1, 1, 3 :: Int)]
+      let (m :: CSR S.Vector Int 3 4 Int) = cooToCsr @U.Vector @S.Vector . fromList $ [(2, 2, -3), (1, 1, 1), (2, 3, 8), (1, 1 :: Int, 3 :: Int)]
       print m
-      m `smIndex` (1, 1) `shouldBe` 4
+      m `csrIndex` (1, 1) `shouldBe` 4
   describe "denseToCSR" $ do
     it ".." $ do
-      let m = denseToCsr (fromList [[(1 :: Int), 0, 2, 0], [0, 4, 0, 0], [0, 0, -3, 8]])
[…]
