+++ src/LatticeSymmetries/Sparse.hs
@@ -1,6 +1,14 @@
-{-# LANGUAGE CApiFFI #-}
-{-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module LatticeSymmetries.Sparse where
 
@@ -8,66 +16,125 @@ module LatticeSymmetries.Sparse where
 
 -- import Data.Vector.Binary
 
-import Control.Exception.Safe (bracket, impureThrow, throwIO)
-import qualified Control.Monad.Primitive as Primitive
+-- import Control.Exception.Safe (bracket, impureThrow, throwIO)
+-- import qualified Control.Monad.Primitive as Primitive
 import Control.Monad.ST
-import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
-import Data.Aeson
-import Data.Aeson.Types (typeMismatch)
+-- import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
+-- import Data.Aeson
+-- import Data.Aeson.Types (typeMismatch)
+
+-- import Data.Scientific (toRealFloat)
+
+-- import Data.Bits (toIntegralSized)
 import Data.Complex
-import qualified Data.List
-import qualified Data.List.NonEmpty as NonEmpty
-import Data.Scientific (toRealFloat)
+-- import qualified Data.List
+-- import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Vector as B
+import qualified Data.Vector.Algorithms.Intro as Intro
+import Data.Vector.Fusion.Bundle (Bundle)
+import qualified Data.Vector.Fusion.Bundle as Bundle (inplace)
+import qualified Data.Vector.Fusion.Bundle.Monadic as Bundle
+import Data.Vector.Fusion.Bundle.Size (Size (..), toMax)
+import Data.Vector.Fusion.Stream.Monadic (Step (..), Stream (..))
+-- import qualified Data.Vector.Fusion.Stream.Monadic as Stream
 import Data.Vector.Generic ((!))
 import qualified Data.Vector.Generic as G
-import qualified Data.Vector.Storable as S
-import qualified Data.Vector.Storable.Mutable as SM
-import qualified Data.Vector.Unboxed as U
-import Data.Yaml (decodeFileWithWarnings)
-import Foreign.C.String (CString, peekCString)
-import Foreign.C.Types (CInt (..), CUInt (..))
-import Foreign.ForeignPtr
-import Foreign.ForeignPtr.Unsafe (unsafeForeignPtrToPtr)
-import Foreign.Marshal.Alloc (alloca, free, malloc, mallocBytes)
-import Foreign.Marshal.Array (newArray, withArrayLen)
-import Foreign.Marshal.Utils (new, with)
-import Foreign.Ptr (FunPtr, Ptr, castPtr)
-import Foreign.StablePtr
-import Foreign.Storable (Storable (..))
+import qualified Data.Vector.Generic.Mutable as GM
+-- import qualified Data.Vector.Storable as S
[…]

[…]
