+++ src/LatticeSymmetries/Sparse.hs
@@ -17,7 +17,7 @@ module LatticeSymmetries.Sparse where
 -- import Data.Vector.Binary
 
 -- import Control.Exception.Safe (bracket, impureThrow, throwIO)
--- import qualified Control.Monad.Primitive as Primitive
+import Control.Monad.Primitive (PrimMonad, PrimState)
 import Control.Monad.ST
 -- import qualified Control.Monad.ST.Unsafe (unsafeIOToST)
 -- import Data.Aeson
@@ -25,10 +25,15 @@ import Control.Monad.ST
 
 -- import Data.Scientific (toRealFloat)
 
--- import Data.Bits (toIntegralSized)
+import Data.Bits (Bits, toIntegralSized)
 import Data.Complex
 -- import qualified Data.List
 -- import qualified Data.List.NonEmpty as NonEmpty
+
+-- import qualified Data.Vector.Fusion.Stream.Monadic as Stream
+
+-- import qualified Data.Vector.Storable as S
+import qualified Data.Vector
 import qualified Data.Vector as B
 import qualified Data.Vector.Algorithms.Intro as Intro
 import Data.Vector.Fusion.Bundle (Bundle)
@@ -36,11 +41,11 @@ import qualified Data.Vector.Fusion.Bundle as Bundle (inplace)
 import qualified Data.Vector.Fusion.Bundle.Monadic as Bundle
 import Data.Vector.Fusion.Bundle.Size (Size (..), toMax)
 import Data.Vector.Fusion.Stream.Monadic (Step (..), Stream (..))
--- import qualified Data.Vector.Fusion.Stream.Monadic as Stream
 import Data.Vector.Generic ((!))
 import qualified Data.Vector.Generic as G
 import qualified Data.Vector.Generic.Mutable as GM
--- import qualified Data.Vector.Storable as S
+import qualified Data.Vector.Storable
+import qualified Data.Vector.Unboxed
 -- import qualified Data.Vector.Storable.Mutable as SM
 -- import Data.Vector.Unboxed (Unbox)
 -- import qualified Data.Vector.Unboxed as U
@@ -89,14 +94,26 @@ natToInt = fromIntegral $ GHC.TypeLits.natVal (Proxy @n)
 data DenseMatrix v (r :: Nat) (c :: Nat) a = DenseMatrix {dmData :: !(v a)}
   deriving stock (Show, Eq, Generic)
 
+type StorableDenseMatrix r c a = DenseMatrix Data.Vector.Storable.Vector r c a
+
+type UnboxedDenseMatrix r c a = DenseMatrix Data.Vector.Unboxed.Vector r c a
+
+type BoxedDenseMatrix r c a = DenseMatrix Data.Vector.Vector r c a
+
+-- | Get number of rows in the matrix
 dmRows :: forall r c a v. KnownDenseMatrix v r c a => DenseMatrix v r c a -> Int
 dmRows _ = natToInt @r
+{-# INLINE dmRows #-}
 
+-- | Get number of columns in the matrix
 dmCols :: forall r c a v. KnownDenseMatrix v r c a => DenseMatrix v r c a -> Int
 dmCols _ = natToInt @c
+{-# INLINE dmCols #-}
 
+-- | Get matrix shape
 dmShape :: forall r c a v. KnownDenseMatrix v r c a => DenseMatrix v r c a -> (Int, Int)
 dmShape m = (dmRows m, dmCols m)
+{-# INLINE dmShape #-}
 
 instance (KnownDenseMatrix v r c a, Num a) => Num (DenseMatrix v r c a) where
[…]
