+++ test/Spec.hs
@@ -1,10 +1,12 @@
 module Main (main) where
 
 import Data.Complex
+import Data.Type.Equality
 import qualified Data.Vector.Storable as S
 import qualified Data.Vector.Unboxed as U
 import Data.Yaml.Aeson
 import Foreign.Storable
+import GHC.Exts (IsList (..))
 import LatticeSymmetries
 import LatticeSymmetries.IO
 import LatticeSymmetries.Parser
@@ -12,6 +14,7 @@ import LatticeSymmetries.Sparse
 import LatticeSymmetries.Types
 import Test.Hspec
 import Text.Parsec (parse)
+import Prelude hiding (toList)
 
 anySpinEDException :: Selector SpinEDException
 anySpinEDException = const True
@@ -180,6 +183,54 @@ main = hspec $ do
       print b
       print c
       a + b `shouldBe` c
+    it "(+)" $ do
+      let a₁' =
+            fromList @(SomeCSR S.Vector Int Double) $
+              [ (0, 0, 0.05052698254161314),
+                (0, 1, 0.34388186562432865),
+                (0, 2, 0.3736125370797053),
+                (0, 4, 0.37422323470438656),
+                (1, 1, 0.08956717735368347),
+                (1, 3, 0.20028803087704727),
+                (2, 0, 0.010520730797154432),
+                (2, 1, 0.7913535778080222),
+                (2, 3, 0.6557777513743904),
+                (2, 4, 0.9609587020099947)
+              ]
+          b₁' =
+            fromList @(SomeCSR S.Vector Int Double) $
+              [ (2, 3, 0.3942046813633675),
+                (2, 4, 0.0008746925379253812)
+              ]
+          (c₁ :: [(Int, Int, Double)]) =
+            [ (0, 0, 0.05052698254161314),
+              (0, 1, 0.34388186562432865),
+              (0, 2, 0.3736125370797053),
+              (0, 4, 0.37422323470438656),
+              (1, 1, 0.08956717735368347),
+              (1, 3, 0.20028803087704727),
+              (2, 0, 0.010520730797154432),
+              (2, 1, 0.7913535778080222),
+              (2, 3, 1.049982432737758),
+              (2, 4, 0.96183339454792)
+            ]
+          a₂ =
+            fromList @(CSR S.Vector Int 5 2 Double) $
+              [(1, 1, 0.9913900269818051), (2, 0, 0.701927418412059)]
+          b₂ =
+            fromList @(CSR S.Vector Int 5 2 Double) $
+              [(2, 0, 0.5563116167541139), (2, 1, 0.5507459794849707), (3, 1, 0.761606773358582)]
+          (c₂ :: [(Int, Int, Double)]) =
+            [ (1, 1, 0.9913900269818051),
+              (2, 0, 1.2582390351661727),
+              (2, 1, 0.5507459794849707),
+              (3, 1, 0.761606773358582)
+            ]
+      withSomeCsr a₁' $ \a₁ -> withSomeCsr b₁' $ \b₁ ->
+        case sameShape a₁ b₁ of
+          Just Refl -> toList (a₁ + b₁) `shouldBe` c₁
+          x -> x `shouldSatisfy` isJust
+      toList (a₂ + b₂) `shouldBe` c₂
   describe "pPrimitiveOperator" $ do
     it ".." $ do
       parse pPrimitiveOperator "" ("c†↓₁₀" :: Text)
@@ -207,28 +258,33 @@ main = hspec $ do
[…]
