### 2022-01-04 db3769b
Implementing operator algebra
