### 2022-01-19 2fce9a0
Simplifying polynomials
