+++ src/LatticeSymmetries/Algebra.hs
@@ -1,115 +1,580 @@
+{-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE OverloadedLists #-}
+{-# LANGUAGE UndecidableInstances #-}
+
 module LatticeSymmetries.Algebra where
 
+import Control.Exception (assert)
 import Control.Monad.ST
 import Data.Complex
+import qualified Data.List as List
+import Data.Ratio ((%))
 import Data.Vector (Vector)
 import qualified Data.Vector
 import qualified Data.Vector.Algorithms.Intro
 import Data.Vector.Generic ((!), (//))
 import qualified Data.Vector.Generic as G
 import qualified Data.Vector.Generic.Mutable as GM
+import GHC.Exts (IsList (..))
 import qualified GHC.Show as GHC
-import LatticeSymmetries.Sparse (combineNeighbors)
+import LatticeSymmetries.Sparse (DenseMatrix, KnownDenseMatrix, combineNeighbors, denseDot, denseMatMul)
+import qualified LatticeSymmetries.Sparse as Sparse
+import Prelude hiding (identity)
 
-data SpinGeneratorType = SpinPlus | SpinMinus | SpinZ
+data SpinGeneratorType = SpinIdentity | SpinZ | SpinPlus | SpinMinus
+  deriving stock (Eq, Ord, Show, Enum, Bounded, Generic)
 
-data FermionGeneratorType = FermionAnnihilate | FermionCreate
+data FermionGeneratorType = FermionIdentity | FermionCount | FermionCreate | FermionAnnihilate
+  deriving stock (Eq, Ord, Show, Enum, Bounded, Generic)
+
+data Generator i g = Generator !i !g
   deriving stock (Eq, Ord, Show, Generic)
 
-data SpinGenerator = SpinGenerator {sgType :: !SpinGeneratorType, sgId :: !Int}
+-- class HasIdentity g where
+--   identity :: g
 
-data FermionGenerator = FermionGenerator {fgId :: !Int, fgType :: !FermionGeneratorType}
-  deriving stock (Eq, Ord, Generic)
+-- instance HasIdentity SpinGeneratorType where
+--   identity = SpinIdentity
 
-instance GHC.Show FermionGenerator where
-  show (FermionGenerator i t) = letter <> dagger <> index
+-- instance HasIdentity FermionGeneratorType where
+--   identity = FermionIdentity
+
+class KnownNat r => HasMatrixRepresentation g r | g -> r where
+  matrixRepresentation :: (G.Vector v c, Num c) => g -> DenseMatrix v r r c
+
+instance HasMatrixRepresentation SpinGeneratorType 2 where
+  matrixRepresentation = spinMatrixRepresentation
+
+instance HasMatrixRepresentation FermionGeneratorType 2 where
+  matrixRepresentation = fermionMatrixRepresentation
+
+spinMatrixRepresentation :: (G.Vector v c, Num c) => SpinGeneratorType -> DenseMatrix v 2 2 c
+spinMatrixRepresentation g = fromList $ case g of
+  SpinIdentity -> [[1, 0], [0, 1]]
+  SpinZ -> [[1, 0], [0, -1]]
+  SpinPlus -> [[0, 1], [0, 0]]
+  SpinMinus -> [[0, 0], [1, 0]]
+
+fermionMatrixRepresentation :: (G.Vector v c, Num c) => FermionGeneratorType -> DenseMatrix v 2 2 c
+fermionMatrixRepresentation g = fromList $ case g of
[…]
