### 2022-01-20 abfa2a3
Combined Algebra with Parser (see Tests); Not sure whether commutators on different sites are computed correctly...
