+++ src/LatticeSymmetries/Parser.hs
@@ -1,18 +1,24 @@
+{-# LANGUAGE OverloadedLists #-}
+
 module LatticeSymmetries.Parser
-  ( pPrimitiveOperator,
+  ( pFermionicOperator,
+    pSpinOperator,
     pOperatorString,
-    PrimitiveOperator (..),
-    SpinOperatorType (..),
     SpinIndex (..),
-    FermionicOperatorType (..),
-    sortByWithParity,
-    Parity (..),
   )
 where
 
 import qualified Data.List.NonEmpty as NonEmpty
+import Data.Ratio
+import qualified Data.Vector.Generic as G
+import LatticeSymmetries.Algebra
+import LatticeSymmetries.ComplexRational
 import Text.Parsec
-import Prelude hiding ((<|>))
+import Prelude hiding (Product, Sum, (<|>))
+
+type ℝ = Rational
+
+type ℂ = ComplexRational
 
 pSubscriptDigit :: Stream s m Char => ParsecT s u m Char
 pSubscriptDigit =
@@ -68,103 +74,122 @@ getSiteIndex (SpinfulFermionicOperator _ _ _ i) = i
 getSiteIndex (SpinlessFermionicOperator _ _ i) = i
 getSiteIndex (SpinOperator _ _ i) = i
 
-pFermionicOperatorType :: Stream s m Char => ParsecT s u m (Char, FermionicOperatorType)
+class KnownIndex i where
+  pIndex :: Stream s m Char => ParsecT s u m i
+
+instance KnownIndex Int where
+  pIndex = pUnicodeSubscriptNumber
+
+instance KnownIndex (SpinIndex, Int) where
+  pIndex = flip (,) <$> pUnicodeSubscriptNumber <*> pSpin
+
+pFermionicOperatorType :: Stream s m Char => ParsecT s u m (Char, FermionGeneratorType)
 pFermionicOperatorType =
   choice
-    [ (,) <$> char 'n' <*> pure FermionicNumberCountingOperator,
-      (,) <$> letter <*> pType
+    [ (,) <$> char 'n' <*> pure FermionCount,
+      (,) <$> char 'c' <*> pType
     ]
   where
     pType =
       isJust <$> optionMaybe (char '†') >>= \x -> case x of
-        True -> pure FermionicCreationOperator
-        False -> pure FermionicAnnihilationOperator
+        True -> pure FermionCreate
+        False -> pure FermionAnnihilate
 
-pFermionicOperator :: Stream s m Char => ParsecT s u m PrimitiveOperator
+pFermionicOperator ::
+  (KnownIndex i, Num c, Stream s m Char) =>
+  ParsecT s u m (Sum (Scaled c (Generator i FermionGeneratorType)))
 pFermionicOperator = do
-  (c, t) <- pFermionicOperatorType
-  s <- optionMaybe pSpin
-  i <- pUnicodeSubscriptNumber
-  pure $ case s of
-    Just s' -> SpinfulFermionicOperator t c s' i
-    Nothing -> SpinlessFermionicOperator t c i
+  (_, t) <- pFermionicOperatorType
+  i <- pIndex
+  pure $ [Scaled 1 (Generator i t)]
 
-pSpinOperator :: Stream s m Char => ParsecT s u m PrimitiveOperator
+pSpinOperator ::
+  Stream s m Char =>
+  ParsecT s u m (Sum (Scaled ComplexRational (Generator Int SpinGeneratorType)))
 pSpinOperator = do
   c <- oneOf "σS" <?> "one of σ, S"
   superscript <- oneOf "⁺⁻ˣʸᶻ" <?> "one of ⁺, ⁻, ˣ, ʸ, ᶻ"
   i <- pUnicodeSubscriptNumber
[…]
