+++ src/LatticeSymmetries/Algebra.hs
@@ -2,8 +2,22 @@
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# LANGUAGE NoMonomorphismRestriction #-}
 
-module LatticeSymmetries.Algebra where
+module LatticeSymmetries.Algebra
+  ( SpinGeneratorType (..),
+    FermionGeneratorType (..),
+    HasMatrixRepresentation (..),
+    Generator (..),
+    Scaled (..),
+    Sum (..),
+    Product (..),
+    scale,
+    -- sumToCanonical,
+    expandProduct,
+    simplify,
+  )
+where
 
 import Control.Exception (assert)
 import Control.Monad.ST
@@ -20,7 +34,7 @@ import GHC.Exts (IsList (..))
 import qualified GHC.Show as GHC
 import LatticeSymmetries.Sparse (DenseMatrix, KnownDenseMatrix, combineNeighbors, denseDot, denseMatMul)
 import qualified LatticeSymmetries.Sparse as Sparse
-import Prelude hiding (identity)
+import Prelude hiding (Product, Sum, identity, toList)
 
 data SpinGeneratorType = SpinIdentity | SpinZ | SpinPlus | SpinMinus
   deriving stock (Eq, Ord, Show, Enum, Bounded, Generic)
@@ -30,6 +44,12 @@ data FermionGeneratorType = FermionIdentity | FermionCount | FermionCreate | Fer
 
 data Generator i g = Generator !i !g
   deriving stock (Eq, Ord, Show, Generic)
+
+data Scaled c g = Scaled !c !g
+  deriving stock (Eq, Ord, Show, Generic)
+
+withoutOrderingCoefficients :: Ord g => Scaled c g -> Scaled c g -> Ordering
+withoutOrderingCoefficients (Scaled _ a) (Scaled _ b) = compare a b
 
 -- class HasIdentity g where
 --   identity :: g
@@ -78,6 +98,26 @@ getBasisExpansion m = filter (\(c, _) -> c /= 0) $ getFactor <$> getValues
           !norm = denseDot a a
        in (dot / norm, g)
 
+getBasisExpansion' ::
+  (Enum g, Bounded g, HasMatrixRepresentation g r, KnownDenseMatrix v r r c, Eq c, Fractional c) =>
+  DenseMatrix v r r c ->
+  Sum (Scaled c g)
+getBasisExpansion' m =
+  Sum
+    . G.map (\(c, g) -> Scaled c g)
+    . G.filter (\(!c, _) -> c /= 0)
+    . G.map getFactor
+    $ G.fromList getValues
+  where
+    getFactor !g =
+      let !a = matrixRepresentation g
+          !dot = denseDot a m
+          !norm = denseDot a a
+       in (dot / norm, g)
+
+isIdentity :: forall g r. HasMatrixRepresentation g r => g -> Bool
+isIdentity g = matrixRepresentation g == (Sparse.denseEye :: DenseMatrix Vector r r Rational)
+
 data CommutatorType = Commutator | Anticommutator
   deriving stock (Show, Eq)
 
@@ -99,12 +139,11 @@ class Ord g => Algebra g where
     g ->
     g ->
     CommutationResult c g
-  commute a b
-    | isDiagonal a || isDiagonal b = CommutationResult []
-    | otherwise = case nonDiagonalCommutatorType @g of
-      Commutator -> CommutationResult commutator
-      Anticommutator -> AnticommutationResult anticommutator
+  commute a b = case nonDiagonalCommutatorType @g of
[…]

[…]
