### 2022-01-20 ffa5a7c
Test products with identities
