### 2022-01-20 af1bc9d
Some more C drafts
