+++ cbits/helpers.h
@@ -2,16 +2,16 @@
 #define HELPERS_H
 
 // #include "LatticeSymmetries_stub.h"
-#include "lattice_symmetries_haskell.h"
+// #include "lattice_symmetries_haskell.h"
 #include <stdint.h>
 
 typedef struct ls_csr_matrix {
-  unsigned dimension;
-  unsigned number_nonzero;
   unsigned *offsets;
   unsigned *columns;
   _Complex double *off_diag_elements;
   _Complex double *diag_elements;
+  unsigned dimension;
+  unsigned number_nonzero;
 } ls_csr_matrix;
 
 typedef struct ls_bit_index {
@@ -19,66 +19,64 @@ typedef struct ls_bit_index {
   uint8_t bit;
 } ls_bit_index;
 
-typedef unsigned (*ls_term_gather_fn)(uint64_t const * /*source*/,
-                                      ls_bit_index const * /*tuple*/);
-typedef void (*ls_term_scatter_fn)(unsigned, ls_bit_index const * /*tuple*/,
-                                   uint64_t * /*destination*/);
-typedef void (*ls_spin_copy_fn)(uint64_t const * /*source*/,
-                                uint64_t * /*destination*/);
-typedef void (*ls_spin_fill_fn)(uint64_t const * /*2D source*/,
-                                uint64_t /*count*/,
-                                uint64_t * /*2D destination*/);
-
+// WARNING: be very careful about allocating ls_term. It needs additional
+// storage!
 typedef struct ls_term {
-  ls_csr_matrix matrix;
-  unsigned number_tuples;
-  unsigned tuple_size;
-  ls_bit_index *tuples;
-  ls_term_gather_fn gather_fn;
-  ls_term_scatter_fn scatter_fn;
+  ls_csr_matrix const *const matrix;
+  unsigned const tuple_size;   // 32 bits
+  ls_bit_index const tuple[2]; // 2 * 16 bits
 } ls_term;
 
-typedef struct ls_sparse_operator {
-  ls_flat_spin_basis const *basis;
-  unsigned number_terms;
-  ls_term *terms;
-} ls_sparse_operator;
+void ls_internal_apply_operator(uint64_t number_terms, ls_term const terms[],
+                                uint64_t number_spins, uint64_t const *spins,
+                                uint64_t capacity, uint64_t number_words,
+                                uint64_t *out_spins,
+                                _Complex double *out_coeffs,
+                                uint64_t *out_counts);
 
-typedef struct ls_output_buffer {
-  uint64_t *spins;
-  _Complex double *coeffs;
-  _Complex double *const diagonal;
-  uint64_t const number_words;
-  ls_spin_copy_fn const spin_copy;
-  ls_spin_fill_fn const spin_fill;
-} ls_output_buffer;
+// typedef unsigned (*ls_term_gather_fn)(uint64_t const * /*source*/,
+//                                       ls_bit_index const * /*tuple*/);
+// typedef void (*ls_term_scatter_fn)(unsigned, ls_bit_index const * /*tuple*/,
+//                                    uint64_t * /*destination*/);
+// typedef void (*ls_spin_copy_fn)(uint64_t const * /*source*/,
+//                                 uint64_t * /*destination*/);
+// typedef void (*ls_spin_fill_fn)(uint64_t const * /*2D source*/,
+//                                 uint64_t /*count*/,
+//                                 uint64_t * /*2D destination*/);
 
-unsigned ls_internal_term_gather_1(uint64_t const *, ls_bit_index const *);
-unsigned ls_internal_term_gather_2(uint64_t const *, ls_bit_index const *);
[…]
