+++ cbits/helpers.c
@@ -1,11 +1,54 @@
 #include "helpers.h"
+#include <assert.h>
 #include <string.h>
 
-void ls_csr_matrix_from_dense(unsigned const dimension,
-                              _Complex double const *const dense,
-                              unsigned *offsets, unsigned *columns,
-                              _Complex double *off_diag_elements,
-                              _Complex double *diag_elements) {
+static inline void ls_internal_copy_spin(int const number_words,
+                                         uint64_t const *const restrict source,
+                                         uint64_t *const restrict destination) {
+  for (int j = 0; j < number_words; ++j) {
+    destination[j] = source[j];
+  }
+}
+
+static inline void ls_internal_fill_with_spin(
+    int const number_words, uint64_t const *const restrict source,
+    uint64_t const count, uint64_t *const restrict destination) {
+  for (uint64_t i = 0; i < count; ++i) {
+    ls_internal_copy_spin(number_words, source, destination + i * number_words);
+  }
+}
+
+static inline unsigned
+ls_internal_gather_bits(int const n, ls_bit_index const *const restrict tuple,
+                        uint64_t const *const restrict source) {
+  unsigned r = 0;
+  for (int i = 0; i < n; ++i) {
+    unsigned const b = (source[tuple[i].word] >> tuple[i].bit) & 1U;
+    // NOTE: Order is important here!
+    // I.e. doing (r | b) << 1U will break stuff!
+    r = (r << 1U) | b;
+  }
+  return r;
+}
+
+static inline void
+ls_internal_scatter_bits(int const n, ls_bit_index const *const restrict tuple,
+                         unsigned r, uint64_t *const restrict destination) {
+  for (int i = n; i-- > 0;) {
+    uint64_t const w = destination[tuple[i].word];
+    uint8_t const b = tuple[i].bit;
+    destination[tuple[i].word] =
+        (w & ~((uint64_t)1 << b)) | ((uint64_t)(r & 1U) << b);
+    r >>= 1U;
+  }
+}
+
+#if 0
+static void ls_csr_matrix_from_dense(unsigned const dimension,
+                                     _Complex double const *const dense,
+                                     unsigned *offsets, unsigned *columns,
+                                     _Complex double *off_diag_elements,
+                                     _Complex double *diag_elements) {
   offsets[0] = 0;
   for (unsigned i = 0; i < dimension; ++i) {
     unsigned nonzero_in_row = 0;
@@ -25,12 +68,11 @@ void ls_csr_matrix_from_dense(unsigned const dimension,
   }
 }
 
-void ls_dense_from_csr_matrix(unsigned const dimension,
-                              unsigned const *const offsets,
-                              unsigned const *const columns,
-                              _Complex double const *const off_diag_elements,
-                              _Complex double const *const diag_elements,
-                              _Complex double *const dense) {
+static void ls_dense_from_csr_matrix(
+    unsigned const dimension, unsigned const *const offsets,
+    unsigned const *const columns,
+    _Complex double const *const off_diag_elements,
+    _Complex double const *const diag_elements, _Complex double *const dense) {
   memset(dense, 0,
          (size_t)dimension * (size_t)dimension * sizeof(_Complex double));
   for (unsigned i = 0; i < dimension; ++i) {
@@ -39,6 +81,78 @@ void ls_dense_from_csr_matrix(unsigned const dimension,
       dense[i * dimension + j] = off_diag_elements[k];
     }
     dense[i * dimension + i] = diag_elements[i];
+  }
+}
+#endif
+
+typedef struct ls_internal_output_buffer {
[…]
